% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_ref_to_comb}
\alias{convert_ref_to_comb}
\title{Convert from reference cell coding to factor reference coding}
\usage{
convert_ref_to_comb(beta_ref)
}
\arguments{
\item{beta_ref}{Matrix of probit coefficients in reference cell coding. (K*q)x1}
}
\value{
Outputs Kxq matrix \code{beta_comb} of probit coefficients using factor
variable coding
}
\description{
\code{convert_ref_to_comb} converts from reference cell coding to a combination of
factor variable and reference cell coding, referred to as factor reference coding
}
\keyword{internal}
