% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_probs}
\alias{convert_to_probs}
\title{Convert from mixture reference coding to P(Y=1|-) conditional probabilities}
\usage{
convert_to_probs(est_xi, glm_form, V_data, cov_name)
}
\arguments{
\item{est_xi}{Matrix of xi parameter estimates. Kxq}

\item{glm_form}{String specifying formula for probit regression, excluding
outcome and latent class. For example, \code{"~ 1"} for the model with only
latent class as covariates. All variables in \code{glm_form} must be found in \code{V_data}.
Do not specify interaction terms for latent class by additional covariates,
as these terms are already included.}

\item{V_data}{Dataframe of additional regression covariates. nxq. Factor
covariates must be converted to factors. If \code{NULL} (default), no additional
covariates are to be included. All variables in \code{glm_form} must
be found in \code{V_data}.}

\item{cov_name}{String vector of length 1 or 2 specifying the key covariate(s)
for which to obtain the outcome probabilities. All covariate names must be
included in \code{glm_form} and \code{V_data}.}
}
\value{
Returns dataframe \code{Phi_df} of the converted outcome probabilities
(i.e., \eqn{\Phi} values) for the categories of the covariate(s) specified in
\code{cov_name}. Number of rows is equal to the number of category combinations of
the key covariate(s). The first K columns include the \eqn{\Phi} values
for the K latent classes, evaluated at the covariate values listed in the
remaining q columns. The key covariate(s) are evaluated at all levels and
the other covariates are evaluated at their reference levels.
}
\description{
Convert regression estimates \eqn{\xi} from mixture reference coding to
conditional probit regression probabilities, P(Y=1|-), for a given covariate.
}
\keyword{internal}
