% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{create_categ_var}
\alias{create_categ_var}
\title{Create a categorical variable using multinomial logistic regression}
\usage{
create_categ_var(beta_mat, design_mat, split_dim = NULL, V = NULL)
}
\arguments{
\item{beta_mat}{Coefficient parameters for the linear predictor terms of a
multinomial logistic regression. KxQ, where K is the number of categories
and Q is the number of covariate terms in the regression}

\item{design_mat}{Regression design matrix. NxQ}

\item{split_dim}{Optional covariate to get separate category probabilities
for. Default is \code{NULL}. If not \code{NULL}, must also specify \code{V} containing the
covariate of interest.}

\item{V}{Optional dataframe of covariate variables. Default is \code{NULL}. If
\code{split_dim} is not \code{NULL}, must be specified as a dataframe of dimension NxQ'
where Q' is the number of covariate variables.}
}
\value{
Returns list \code{out_vars} containing
\describe{
\item{\code{categ_var}}{Generated categorical variable. Nx1}
\item{\code{pi_mat}}{Matrix of category probabilities for all individuals. NxK}
\item{\code{true_pi}}{Vector of overall category probabilities across all
individuals. Nx1}
\item{\code{pi_split}}{If \code{split_dim} is specified, HxK matrix of category
probabilities split by the levels in \code{split_dim}, where H is the number of
levels in the \code{split_dim} variable, and K is the number of categories as
specified by \code{beta_mat}. Otherwise, \code{NULL} if \code{split_dim} is \code{NULL}.}
}
}
\description{
Create a categorical variable using multinomial logistic regression
}
\examples{
# Define multinomial logistic regression formula to depend on s_all
formula_c <- "~ s_all"
# Create dataframe of covariate variables
V <- data.frame(s_all = as.factor(c(rep(0, times = 60000), 
                                    rep(1, times = 20000))))
# Create design matrix
design_mat_c <- model.matrix(as.formula(formula_c), data = V) 
# Specify matrix of multinomial logistic regression coefficients for each
# category. Kx2 matrix with first row all 0's.
# The coefficients below corresponds to category probabilities of 
# (0.3, 0.5, 0.2) for S=1 and (0.1, 0.6, 0.3) for S=2, giving overall 
# probabilities (0.253, 0.522, 0.225)
beta_mat_c <- matrix(c(0, 0, 
                       0.5, 1.3,
                       -0.4, 1.5), byrow = TRUE, nrow = 3, ncol = 2)
# Create categorical variable                      
out_vars <- create_categ_var(beta_mat = beta_mat_c, design_mat = design_mat_c,
                             split_dim = "s_all", V = V)
out_vars$categ_var
}
\seealso{
\code{\link[=simulate_pop]{simulate_pop()}}
}
