% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{create_traceplot}
\alias{create_traceplot}
\title{Create traceplots}
\usage{
create_traceplot(param_mcmc, param_names)
}
\arguments{
\item{param_mcmc}{List output from \code{\link[=get_param_mcmc]{get_param_mcmc()}} containing:
\describe{
\item{\code{pi_mcmc}}{MxK dataframe of the class membership probability
parameters, \eqn{\pi}, where M is the number of iterations and K is the
number of latent classes.}
\item{\code{theta_mcmc}}{Mx(JxKxR) dataframe of the item level probability
parameters, \eqn{\theta}, where J is the number of items and R is the
maximum number of item levels.}
\item{\code{xi_mcmc}}{If output for a \code{swolca} object, Mx(Kxq) dataframe of
the regression parameters, \eqn{\xi}, where q is the number of covariates,
excluding latent class indicators, in the regression.}
}}

\item{param_names}{String vector of parameter names to create plots for. All
names must be found as column names of the dataframes in \code{param_mcmc}.}
}
\value{
Creates a grid of traceplots for the specified parameters
}
\description{
Create traceplots
}
\examples{
data(run_nhanes_swolca_results)
param_mcmc <- get_param_mcmc(res = run_nhanes_swolca_results)
# Specify selection of pi_1 to pi_5
param_names <- colnames(param_mcmc$pi_mcmc)
# Create traceplots
create_traceplot(param_mcmc = param_mcmc, param_names = param_names)

}
\seealso{
\code{\link[=get_param_mcmc]{get_param_mcmc()}} \link{create_acfplot}
}
