% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_adjust.R
\name{DEadj}
\alias{DEadj}
\title{Adjust estimates}
\usage{
DEadj(par, par_hat, R2R1)
}
\arguments{
\item{par}{Unadjusted parameter estimates}

\item{par_hat}{Unadjusted median parameter estimates}

\item{R2R1}{Adjustment matrix}
}
\value{
Outputs adjusted parameter estimates
}
\description{
`DEadj` is a helper function to apply the matrix rotation
}
\keyword{internal}
