% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{get_categ_probs}
\alias{get_categ_probs}
\title{Get unique categorical probabilities for multinomial logistic regresion}
\usage{
get_categ_probs(beta_mat, formula, V)
}
\arguments{
\item{beta_mat}{Coefficient parameters for the linear predictor terms of a
multinomial logistic regression. KxQ, where Q is the number of covariate
terms in the regression}

\item{formula}{Formula for multinomial logistic regression. Should start with
\code{"~ c_all"} if generating exposure X. All variables must be found in \code{V}.}

\item{V}{Dataframe of variables}
}
\value{
List \code{categ_probs} containing:
\describe{
\item{\code{unique_comb}}{Matrix of unique covariate combinations in the
design matrix created using the specified formula. mxQ, where m is the
number of unique covariate combinations.}
\item{\code{unique_prob_mat}}{Matrix of category probabilities for all
unique covariate cominations. mxQ}
}
}
\description{
Get categorical probabilities for a unique design matrix corresponding to
specified multinomial logistic regression formula and coefficient parameters
}
\examples{
# Get pi probabilities for C ~ S
V_unique <- data.frame(s_all = as.factor(1:2))
formula_lc <- "~ s_all"
beta_mat <- matrix(c(0, 0,  # Kx2 (first row is all 0's)
                     0.5, 1.3,
                     -0.4, 1.5), byrow = TRUE, nrow = 3, ncol = 2)
categ_probs_pi <- get_categ_probs(beta_mat = beta_mat, formula = formula_lc, 
                               V = V_unique)

# Get theta probabilities for X ~ C + S
V_unique <- expand.grid(factor(1:3), factor(1:2))  
colnames(V_unique) <- c("c_all", "s_all")    
formula <- "~ c_all + s_all"
non_div_mode <- log(0.05 / 0.85)
mode_div_non <- log(0.85 / 0.05)
# r=1: 0.85, 0.05, 0.05; None, High, Med
beta_mat <- matrix(c(0,            0,              0,              0,
                     non_div_mode, mode_div_non,   mode_div_non,   0,
                     non_div_mode, mode_div_non,   2*mode_div_non, 0,
                     non_div_mode, 2*mode_div_non, mode_div_non,   0),
                   nrow = 4, byrow = TRUE)
categ_probs_theta <- get_categ_probs(beta_mat = beta_mat, formula = formula, 
                                     V = V_unique)

# Add in association with S where categ High has prob 0.91 for S=1
beta_mat[-1, 4] <- c(0, 0, log(0.91/0.03) - mode_div_non)
categ_probs_theta_s <- get_categ_probs(beta_mat = beta_mat, formula = formula, 
                                       V = V_unique)

}
\keyword{internal}
