% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ci}
\alias{get_ci}
\title{Get confidence or credible interval}
\usage{
get_ci(
  post_samp,
  quant_lb = 0.025,
  quant_ub = 0.975,
  digits = 2,
  string = FALSE
)
}
\arguments{
\item{post_samp}{Numeric vector of posterior samples}

\item{quant_lb}{Lower bound quantile of interval estimate. Default is 0.025
corresponding to a 95\% interval}

\item{quant_ub}{Upper bound quantile of interval estimate. Default is 0.975
corresponding to a 95\% interval}

\item{digits}{Number of digits to round to. Default is 2.}

\item{string}{Boolean indicating if the interval should be returned as a
string with parentheses. Default is \code{FALSE}, which returns the interval as a
vector.}
}
\value{
Outputs string \code{ci} with formatted interval
}
\description{
\code{get_ci} is a helper function that formats confidence or credible intervals
}
\keyword{internal}
