% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_cov_props}
\alias{get_cov_props}
\title{Get column percentages across latent classes}
\usage{
get_cov_props(
  svy_design,
  cov,
  var_levels,
  col_props = TRUE,
  res = NULL,
  digits = 1
)
}
\arguments{
\item{svy_design}{Survey design defined using the \code{svydesign()} function in
the \code{survey} package}

\item{cov}{String specifying covariate of interest. Use \code{"population"} or
\code{"sample"} to get the class-specific population or sample sizes, respectively.}

\item{var_levels}{String vector specifying levels of the covariate of interest.
If the covariate is continuous, set this to the empty string \code{""}.}

\item{col_props}{Boolean indicating if factor variables should have percentages
reported as column totals (default \code{TRUE}) that provide the percentage of the
population in each category for a given latent class, or reported as row
totals that provide the percentage of the population in each latent class for
a given category.}

\item{res}{Results from unadjusted or adjusted SWOLCA or WOLCA, necessary for
\verb{cov == "population}. Otherwise, use default value of \code{NULL}.}

\item{digits}{Integer indicating the number of decimal places to be used.
Default is 1, which rounds to the nearest tenth.}
}
\value{
Outputs \code{output} dataframe with number of rows equal to the number of levels
in the covariate of interest (one row if covariate is continuous), and number
of columns equal to the number of latent classes, plus an additional column
for the overall value over all classes.
}
\description{
\code{get_cov_props} gives the column means or percentages across latent classes
for a given demographic variable, calculated for the population using survey
weights and methods from the \code{survey} package
}
\details{
If the covariate is a factor, outputs percentages for each level. If the
covariate is continuous, outputs the mean. If the covariate is the population
class membership proportions (i.e., \code{cov == "population"}), the posterior
standard deviation is also provided using the results from SWOLCA or WOLCA.
The standard errors for the other variables are not provided because they
would be underestimates due to not accounting for variability in the latent
class assignments. If the covariate is the sample class membership proportions
(i.e., \code{cov == "sample"}), then the percentages provided are exact.
}
\keyword{internal}
