% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_dic}
\alias{get_dic}
\title{Obtain DIC-6}
\usage{
get_dic(res)
}
\arguments{
\item{res}{An object of class \code{"swolca"} or \code{"wolca"}, resulting from a call
to \code{\link[=swolca]{swolca()}}, \code{\link[=swolca_var_adjust]{swolca_var_adjust()}}, \code{\link[=wolca]{wolca()}}, or \code{\link[=wolca_var_adjust]{wolca_var_adjust()}}.}
}
\value{
Returns the DIC-6.
}
\description{
\code{get_dic} returns the DIC-6 for the model to be used as a measure of model fit.
}
\details{
The deviance information criterion (DIC) can be used as a metric for model
goodness of fit for mixture models (Spiegelhater et al., 2002; Celeux et al., 2006).
It compares two values:
\enumerate{
\item the median of the likelihood over all iterations
\item the likelihood calculated with the posterior median estimates of all parameters.
We use the DIC-6, which increases the penalty of complexity to reduce
overfitting (Stephenson et al., 2022). The DIC-6 is calculated as
\eqn{3\bar{D}(\theta)-2D(\bar{\theta})) 
= -6E[\log L(D|\theta)|D] + 4\log L(y|\bar{\theta})}.
where \eqn{E[\cdot]} denotes the expectation, \eqn{\bar{D}(\theta)} is the
posterior median observed deviance and \eqn{D(\bar{\theta})} is the deviance
of the posterior median.
}
}
\examples{
data(run_nhanes_swolca_results)
dic <- get_dic(res = run_nhanes_swolca_results)

}
\references{
Celeux, G., Forbes, F., Robert, C. P., Titterington, D. M. et al. (2006)
Deviance information criteria for missing data models. Bayesian Analysis, 1, 651-673.

Spiegelhalter, D. J., Best, N. G., Carlin, B. P. and Van Der Linde, A. (2002)
Bayesian measures of model complexity and fit. Journal of the Royal
Statistical Society: Series B (Statistical Methodology), 64, 583-639.

Stephenson, B. J., Herring, A. H., and Olshan, A. F. (2022). Derivation of
maternal dietary patterns accounting for regional heterogeneity. Journal of
the Royal Statistical Society Series C: Applied Statistics 71, 1957–1977.
}
\seealso{
\code{\link[=summarize_res]{summarize_res()}}
}
