% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Get posterior estimates}
\usage{
get_estimates(MCMC_out, post_MCMC_out, n, p, V, y_all, x_mat)
}
\arguments{
\item{MCMC_out}{Output from `run_MCMC_Rcpp` containing `pi_MCMC`, 
`theta_MCMC`, `xi_MCMC`, `c_all_MCMC`, `z_all_MCMC`, and `loglik_MCMC`}

\item{post_MCMC_out}{output from `post_process` containing `K_med`, `pi`, 
`theta`, `xi`}

\item{n}{Number of individuals}

\item{V}{Regression design matrix without class assignment. nxq}

\item{y_all}{Vector of outcomes. nx1}

\item{x_mat}{Categorical exposure matrix. nxp}
}
\value{
Returns list `estimates` containing:
\describe{
  \item{\code{K_red}}{Number of unique classes}
  \item{\code{pi_red}}{Matrix of final posterior samples for pi. Mx(K_red)}
  \item{\code{theta_red}}{Array of final posterior samples for theta. Mxpx(K_red)xd}
  \item{\code{xi_red}}{Array of final posterior samples for xi. Mx(K_red)xq}
  \item{\code{pi_med}}{Vector of posterior median estimates for pi. (K_red)x1}
  \item{\code{theta_med}}{Array of posterior median estimates for theta. px(K_red)xd}
  \item{\code{xi_med}}{Matrix of posterior median estimates for xi. (K_red)xq}
  \item{\code{Phi_med}}{Vector of final individual outcome probabilities. nx1}
  \item{\code{c_all}}{Vector of final individual class assignments. nx1}
  \item{\code{pred_class_probs}}{Matrix of individual posterior class probabilities. nx(K_red)}
  \item{\code{loglik_med}}{Vector of final indiviudal log-likehoods. nx1} 
}
}
\description{
`get_estimates` obtains posterior parameter samples and estimates prior to
variance adjustment
}
