% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_levels}
\alias{get_levels}
\title{Get levels of a variable}
\usage{
get_levels(df, var)
}
\arguments{
\item{df}{Dataframe including variable of interest}

\item{var}{String specifying variable of interest}
}
\value{
Outputs string vector \code{levels} containing the levels of the variable
of interest. If the variable is not a factor, outputs the empty string \code{""}.
}
\description{
\code{get_levels} obtains the levels of a specified variable in a dataframe. For
factor varialbes, the levels are directly returned. For continuous variables,
the levels are returned as an empty string. Internal function used in
\code{\link[=vars_across_class]{vars_across_class()}}.
}
\seealso{
\code{\link[=vars_across_class]{vars_across_class()}}
}
\keyword{internal}
