% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_mode}
\alias{get_mode}
\title{Get mode}
\usage{
get_mode(v)
}
\arguments{
\item{v}{Input vector}
}
\value{
Outputs most common value found in input vector \code{v}
}
\description{
\code{get_mode} is a helper function that obtains the modal value given an input
vector.
}
\keyword{internal}
