% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regr_coefs.R
\name{get_prob_pos}
\alias{get_prob_pos}
\title{Get posterior probability}
\usage{
get_prob_pos(post_samp, cutoff = 0, digits = 2)
}
\arguments{
\item{post_samp}{Numeric vector of posterior samples}

\item{cutoff}{Value specifying cutoff to compare posterior samples to. Default
is 0.}

\item{digits}{Number of digits to round to. Default is 2, where posterior
probabilities smaller than 0.01 are denoted as "<0.01".}
}
\value{
Outputs \code{prob_pos} posterior probability
}
\description{
\code{get_prob_pos} is a helper function that obtains the probability that the
posterior sample estimates are greater than some specified value
}
\keyword{internal}
