% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_regr_coefs}
\alias{get_regr_coefs}
\title{Obtains table of regression coefficients}
\usage{
get_regr_coefs(res, ci_level = 0.95, digits = 2)
}
\arguments{
\item{res}{An object of class \code{"swolca"} or \code{"wolca"}, resulting from a call
to \code{\link[=swolca]{swolca()}}, \code{\link[=swolca_var_adjust]{swolca_var_adjust()}}, \code{\link[=wolca]{wolca()}}, or \code{\link[=wolca_var_adjust]{wolca_var_adjust()}}.}

\item{ci_level}{Numeric from 0 to 1 specifying the credible interval level.
Default is 0.95, which gives a 95\\% equal-tailed interval composed of the
2.5\\% and 97.5\\% quantiles. For \code{wolca()} results, this must match the
\code{ci_level} parameter in the main function.}

\item{digits}{Integer indicating the number of decimal places to be used.
Default is 2, which rounds to the nearest hundredth.}
}
\value{
Returns dataframe \code{beta} containing the following columns:
\describe{
\item{\code{Covariate}}{Names of covariate terms.}
\item{\code{Estimate}}{Regression estimates in reference cell coding format.}
\item{\code{LB}}{Regression estimate lower bound corresponding to the level
specified in \code{ci_level}. Calculated using posterior samples for \code{swolca}
objects and confidence intervals for \code{wolca} objects.}
\item{\code{UB}}{Regression estimate upper bound corresponding to the level
specified in \code{ci_level}. Calculated using posterior samples for \code{swolca}
objects and confidence intervals for \code{wolca} objects.}
\item{\code{P(xi>0)} or \code{p-value}}{For \code{swolca} objects, the proportion
of posterior samples greater than 0. For \code{wolca} objects, the p-value
obtained from the \code{wolca_svyglm()} function}
}
}
\description{
\code{get_regr_coefs} produces a summary table of the regression coefficients,
converted to standard reference cell coding.
}
\details{
If \code{res} is a \code{swolca} object, any cluster can be chosen as the reference
cluster level for which to display regression coefficients. Simply run
\code{reorder_clusters()} with the desired reference level as the first cluster in
\code{new_order}, and then run \code{get_regr_coefs()} using the reordered \code{res} object.

If \code{res} is a \code{wolca} object, choosing a different reference level can only
be done by re-running \code{wolca_syvglm()} with the new ordering. To do this, run
\code{reorder_clusters()}, then run \code{wolca_svyglm()} to obtain regression
estimates with a different reference class, and then finally use the
\code{get_regr_coefs()} function.
}
\examples{
data(run_nhanes_swolca_results)
get_regr_coefs(res = run_nhanes_swolca_results, ci_level = 0.95, digits = 2)

}
\seealso{
\code{\link[=reorder_clusters]{reorder_clusters()}} \code{\link[=plot_outcome_probs]{plot_outcome_probs()}} \code{\link[=summarize_res]{summarize_res()}}
}
