% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_adjust.R
\name{grad_par}
\alias{grad_par}
\title{Get gradient of log posterior from the unconstrained parameter space}
\usage{
grad_par(pwts, svydata, stan_mod, stan_data, par_stan, u_pars)
}
\arguments{
\item{pwts}{Replicate weights from `svyrepdesign` object}

\item{svydata}{Data frame containing all variables from `svyrepdesign` object}

\item{stan_mod}{Stan model object}

\item{stan_data}{Stan data input}

\item{par_stan}{Parameters with respect to which gradient should be computed}

\item{u_pars}{Unconstrained parameters estimates for evaluating gradient}
}
\value{
Outputs `gradpar` gradient evaluated at `u_pars` using replicate weights
}
\description{
`grad_par` is a helper function nested in `withReplicates()` to obtain the 
gradient with the replicate weights
}
\keyword{internal}
