% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_OLCA.R
\name{init_OLCA}
\alias{init_OLCA}
\title{Initialize OLCA model}
\usage{
init_OLCA(K, n, p, d, alpha, eta)
}
\arguments{
\item{K}{Number of classes}

\item{n}{Number of individuals}

\item{p}{Number of exposure items}

\item{d}{Number of exposure categories}

\item{alpha}{Vector of hyperparameters for pi. Kx1}

\item{eta}{Vector of hyperparameters for theta. dx1}
}
\value{
Returns list `OLCA_params` containing:
\describe{
  \item{\code{pi}}{Vector parameter pi for class membership probabilities. Kx1}
  \item{\code{theta}}{Array parameter theta for item category probabilities. pxKxd}
  \item{\code{c_all}}{Vector of random initial class assignments. nx1}
}
}
\description{
`init_OLCA` initializes priors and variables for an overfitted latent class
analysis (OLCA) given hyperparameters.
}
\examples{
K <- 30; n <- 400; p <- 30; d <- 4
alpha <- rep(1, K) / K
eta <- rep(1, d)
init_OLCA(K = K, n = n, p = p, d = d, alpha = alpha, eta = eta)

}
