% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_probit.R
\name{init_probit}
\alias{init_probit}
\title{Initialize probit model}
\usage{
init_probit(K, n, q, V, mu0, Sig0, y_all, c_all)
}
\arguments{
\item{K}{Number of classes}

\item{n}{Number of individuals}

\item{q}{Number of regression covariates excluding class assignment}

\item{V}{Regression design matrix without class assignment. nxq}

\item{mu0}{List of vectors of mean hyperparameters for xi. List of K qx1 vectors}

\item{Sig0}{List of matrices of variance hyperparameters for xi. List of K qxq matrices}

\item{y_all}{Vector of outcomes. nx1}

\item{c_all}{Vector of random initial class assignments. nx1}
}
\value{
Returns list `probit_params` containing:
\describe{
  \item{\code{xi}}{Matrix parameter xi for probit regression coefficients. Kxq}
  \item{\code{z_all}}{Vector of latent variables in the probit model. nx1}
}
}
\description{
`init_probit` initializes priors and variables for the probit regression model 
given hyperparameters.
}
\examples{
# Probit model only includes latent class 
K <- 30; n <- 400; q <- 1
V <- matrix(1, nrow = n)
mu0 <- Sig0 <- vector("list", K_max)
for (k in 1:K_max) {
  # MVN(0,1) hyperprior for prior mean of xi
  mu0[[k]] <- rnorm(n = q)
  # InvGamma(3.5, 6.25) hyperprior for prior variance of xi. Assume uncorrelated
  # components and mean variance 2.5 for a weakly informative prior on xi
  Sig0[[k]] <- diag(rinvgamma(n = q, shape = 3.5, scale = 6.25), nrow = q, ncol = q)
}
y_all <- rbinom(n = n, size = 1, prob = 0.5)
c_all <- c(rep(1, 100), rep(2, 100), rep(3, 100), rep(4, 100))
init_probit(K = K, n = n, q = q, V = V, mu0 = mu0, Sig0 = Sig0, 
y_all = y_all, c_all = c_all)

}
