% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{manual_CI}
\alias{manual_CI}
\title{Get manual confidence interval}
\usage{
manual_CI(model_object, svy_df, ci = 0.95)
}
\arguments{
\item{model_object}{svyglm model fit object}

\item{svy_df}{survey design df}

\item{ci}{confidence interval level}
}
\value{
Outputs dataframe of confidence interval for all coefficients
}
\description{
`manual_CI` manually calculates a Wald confidence interval using a t-dist
with df from the survey design. Used for WOLCA in the situation where svyglm 
produces negative residual df, calculated as design df plus one, minus the 
number of parameters estimated. Best if no cluster-level covariates in the 
regression model
}
\keyword{internal}
