% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnorm_cpp}
\alias{mvrnorm_cpp}
\title{Draw from a Multivariate Normal distribution}
\usage{
mvrnorm_cpp(n, mu, sigma)
}
\arguments{
\item{n}{Number of draws}

\item{mu}{Vector of means with length equal to number of dimensions}

\item{sigma}{Positive-definite symmetric covariance matrix}
}
\value{
Matrix result of draws from the specified Multivariate Normal 
distribution, with rows corresponding to draws and columns corresponding to 
dimensions.
}
\description{
Draw from a Multivariate Normal distribution
}
\keyword{internal}
