% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_class_dist}
\alias{plot_class_dist}
\title{Plot distribution of classes in the population across posterior sample iterations}
\usage{
plot_class_dist(
  res,
  class_labels = NULL,
  class_title = "Dietary Pattern",
  y_title = "Class Membership Probability",
  ...
)
}
\arguments{
\item{res}{An object of class \code{"swolca"} or \code{"wolca"}, resulting from a call
to \code{\link[=swolca]{swolca()}}, \code{\link[=swolca_var_adjust]{swolca_var_adjust()}}, \code{\link[=wolca]{wolca()}}, or \code{\link[=wolca_var_adjust]{wolca_var_adjust()}}.}

\item{class_labels}{String vector of names for the latent classes. Kx1.
If \code{NULL} (default), numbers from 1 to K are used, where K is the final
determined number of latent classes.}

\item{class_title}{String specifying the title for the latent classes.
Default is \code{"Dietary Pattern"}.}

\item{y_title}{String specifying the title for the y-axis. Default is
\code{"Class Membership Probability"}.}

\item{\dots}{Additional arguments passed}
}
\value{
Returns a \code{ggplot2} object displaying a boxplot of the distribution of class
membership probabilities in the posterior samples, for each latent class.
}
\description{
\code{plot_class_dist} plots a boxplot of the class membership probabilities,
\eqn{\pi}, in the posterior samples, for each latent class.
}
\examples{
data(run_nhanes_swolca_results)

# Default labels
plot_class_dist(res = run_nhanes_swolca_results)

# Specifying labels
class_labels <- paste0("Class ", 1:5)
plot_class_dist(res = run_nhanes_swolca_results, class_labels = class_labels)
}
\seealso{
\code{\link[=plot_pattern_probs]{plot_pattern_probs()}} \code{\link[=plot_pattern_profiles]{plot_pattern_profiles()}} \code{\link[=plot_outcome_probs]{plot_outcome_probs()}}
}
