% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_outcome_probs}
\alias{plot_outcome_probs}
\title{Plot conditional probability of outcome for each latent class for one or two
categorical covariates.}
\usage{
plot_outcome_probs(
  res,
  cov_name,
  ci_level = 0.95,
  add_lines = FALSE,
  cov_labels = NULL,
  class_labels = NULL,
  class_title = "Dietary Pattern",
  x_title = NULL,
  y_title = "Probability of Outcome",
  ...
)
}
\arguments{
\item{res}{An object of class \code{"swolca"} or \code{"wolca"}, resulting from a call
to \code{\link[=swolca]{swolca()}}, \code{\link[=swolca_var_adjust]{swolca_var_adjust()}}, \code{\link[=wolca]{wolca()}}, or \code{\link[=wolca_var_adjust]{wolca_var_adjust()}}.}

\item{cov_name}{String specifying the covariate to plot. To plot
interactions between two covariates, specify a string vector with both
covariate names. Covariates must be included in \code{glm_form} and \code{V_data}.
If no additional covariates are used in the model, please set \code{cov_name = NULL}.}

\item{ci_level}{Optional number from 0 to 1 specifying the confidence/credible
interval level. Default is \code{0.95}, which gives a 95\\% equal-tailed interval
composed of the 2.5\\% and 97.5\\% quantiles. For \code{"wolca"} objects, this must
match the \code{ci_level} parameter in the main \code{\link[=wolca]{wolca()}} function. Set to \code{NULL}
to exclude credible intervals from the plot.}

\item{add_lines}{Boolean specifying whether lines should be added to connect
the points between categories. Default is \code{FALSE}.}

\item{cov_labels}{Optional list of 1 or 2 string vectors specifying the
corresponding category labels for the covariate(s) of interest. Must be the
same length as \code{cov_name} and have the appropriate number of categories.
If \code{NULL} (default), the covariate categories from the data are used.}

\item{class_labels}{String vector of names for the latent classes. Kx1.
If \code{NULL} (default), numbers from 1 to K are used, where K is the final
determined number of latent classes.}

\item{class_title}{String specifying the title for the latent classes.
Default is \code{"Dietary Pattern"}.}

\item{x_title}{Optional string specifying x-axix label. If \code{NULL} (default),
the first string in \code{cov_name}is used.}

\item{y_title}{Optional string specifying the title for the y-axis. Default is
\code{"Probability of Outcome"}.}

\item{\dots}{Additional arguments passed}
}
\value{
Returns a \code{ggplot2} object displaying the point estimates and error bars of
the conditional probability of the outcome across categories of one or two
covariates of interest, colored by latent class.
}
\description{
\code{plot_outcome_probs} plots the conditional probability of the outcome,
ranging from 0 to 1 and obtained by transforming the probit regression
coefficients to the probability scale, for up to two categorical covariates.
Point estimates and error bars are colored by latent class.
}
\examples{
data(run_nhanes_swolca_results)
res <- run_nhanes_swolca_results

# Default labels
plot_outcome_probs(res = res, cov_name = "racethnic")

# Specify labels
cov_labels <- c("NH White", "NH Black", "NH Asian", "Hispanic/Latino", 
                "Other/Mixed")
class_labels <- paste0("Class", 1:5)
x_title <- "Race and Ethnicity"
plot_outcome_probs(res = res, cov_name = "racethnic", cov_labels = cov_labels, 
                   class_labels = class_labels, x_title = x_title)

# Two covariates, remove error bars, add lines
age_cat_categs <- c("(20,40)", "(40,60)", ">=60")
racethnic_categs <- c("NH White", "NH Black", "NH Asian", "Hispanic/Latino", 
                      "Other/Mixed")
plot_outcome_probs(res = res, cov_name = c("age_cat", "racethnic"),
                   cov_labels = list(age_cat_categs, racethnic_categs), 
                   class_labels = class_labels, x_title = "Age Group", 
                   ci_level = NULL, add_lines = TRUE)         
\dontrun{                    
# Multiple plots for various covariates
educ_categs <- c("Some College", "HS/GED", "<HS")
smoker_categs <- c("Non-Smoker", "Smoker")
physactive_categs <- c("Inactive", "Active")
p1 <- plot_outcome_probs(res = res, cov_name = "age_cat",
                         cov_labels = age_cat_categs, 
                         class_labels = class_labels, 
                         x_title = "Age Group")
p2 <- plot_outcome_probs(res = res, cov_name = "racethnic",
                         cov_labels = racethnic_categs, 
                         class_labels = class_labels, 
                         x_title = "Race and Ethnicity")
p3 <- plot_outcome_probs(res = res, cov_name = "smoker",
                         cov_labels = smoker_categs, 
                         class_labels = class_labels, 
                         x_title = "Current Smoking Status")
p4 <- plot_outcome_probs(res = res, cov_name = "physactive",
                         cov_labels = physactive_categs, 
                         class_labels = class_labels, 
                         x_title = "Physical Activity")
ggpubr::ggarrange(p1, p2, p3, p4, common.legend = TRUE, legend = "top", 
                  nrow = 1, ncol = 4, widths = c(0.7, 1, 0.45, 0.45)) 
}                              
                   
}
\seealso{
\code{\link[=get_regr_coefs]{get_regr_coefs()}} \code{\link[=plot_pattern_probs]{plot_pattern_probs()}} \code{\link[=plot_class_dist]{plot_class_dist()}} \code{\link[=plot_pattern_profiles]{plot_pattern_profiles()}}
}
