% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_pi_boxplots}
\alias{plot_pi_boxplots}
\title{Plot pi boxplots across posterior samplers for each latent class}
\usage{
plot_pi_boxplots(
  res,
  class_labels = NULL,
  class_title = "Dietary Pattern",
  y_title = "Class Membership Probability",
  ...
)
}
\arguments{
\item{res}{Output from \code{swolca()}, \code{solca()}, or \code{wolca()} containing at
least \code{estimates} (or \code{estimates_unadj}) and \code{data_vars}}

\item{class_labels}{String vector of names for the latent classes. Kx1.
If \code{NULL} (default), numbers from 1 to K are used, where K is the final
determined number of latent classes.}

\item{class_title}{String specifying the title for the latent classes.
Default is \code{"Dietary Pattern"}.}

\item{y_title}{String specifying the title for the y-axis. Default is
\code{"Class Membership Probability"}.}

\item{\dots}{Additional arguments passed}
}
\value{
Returns a \code{ggplot2} object displaying a boxplot of the distribution of class
membership probabilities in the posterior samples, for each latent class.
}
\description{
\code{plot_pi_boxplots} plots a boxplot of the class membership probabilities,
\eqn{\pi}, in the posterior samples, for each latent class.
}
\examples{
data(run_nhanes_swolca_results)

# Default labels
plot_pi_boxplots(res = run_nhanes_swolca_results)

# Specifying labels
class_labels <- paste0("Class ", 1:5)
plot_pi_boxplots(res = run_nhanes_swolca_results, class_labels = class_labels)
}
\seealso{
\code{\link[=plot_theta_probs]{plot_theta_probs()}} \code{\link[=plot_theta_modes]{plot_theta_modes()}} \code{\link[=plot_Phi_line]{plot_Phi_line()}}
}
