% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_regr_coefs}
\alias{plot_regr_coefs}
\title{Plot regression coefficients}
\usage{
plot_regr_coefs(regr_coefs, res, cov_labels = NULL, ...)
}
\arguments{
\item{regr_coefs}{Dataframe output from \code{\link[=get_regr_coefs]{get_regr_coefs()}} containing at
least the following columns:
\describe{
\item{\code{Covariate}}{Names of covariate terms.}
\item{\code{Estimate}}{Regression estimates in reference cell coding format.}
\item{\code{LB}}{Regression estimate lower bound}
\item{\code{UB}}{Regression estimate upper bound}
}}

\item{res}{An object of class \code{"swolca"} or \code{"wolca"}, resulting from a call
to \code{\link[=swolca]{swolca()}}, \code{\link[=swolca_var_adjust]{swolca_var_adjust()}}, \code{\link[=wolca]{wolca()}}, or \code{\link[=wolca_var_adjust]{wolca_var_adjust()}}.}

\item{cov_labels}{Optional string vector of labels whose order must
correspond exactly with that of the \code{Covariate} column in \code{regr_coefs}.
Default is \code{NULL} and original covariate names are used.}

\item{\dots}{Additional arguments passed}
}
\value{
Returns a \code{ggplot2} object displaying point estimates and error bars for the
probit regression coefficients in reference cell coding format. Estimates and
intervals are obtained from the \code{\link[=get_regr_coefs]{get_regr_coefs()}} function using the
interval level specified there.
}
\description{
\code{plot_regr_coefs} plots point estimates and error bars for all the regression
coefficients in reference cell coding format, using the output from the
\code{\link[=get_regr_coefs]{get_regr_coefs()}} function. Estimates corresponding to the same latent
class are displayed with the same color.
}
\examples{
data(run_nhanes_swolca_results)
res <- run_nhanes_swolca_results

# Get table of regression coefficients
regr_coefs <- get_regr_coefs(res = res, ci_level = 0.95, digits = 2)

# Define new vector of covariate labels
class_dummies <- paste0("C", 2:5)
reps <- length(class_dummies)
age_dummies <- paste0("Age", c("40_60", "60"))
race_dummies <- paste0("RaceEth", c("NH_Black", "NH_Asian", "Hisp", "Other"))
cov_labels <- c("Intercept", class_dummies, age_dummies, race_dummies,
                "SmokerYes", "PhysActive",
                paste0(class_dummies, ":", rep(age_dummies, each = reps)),
                paste0(class_dummies, ":", rep(race_dummies, each = reps)),
                paste0(class_dummies, ":", rep("SmokerYes", each = reps)),
                paste0(class_dummies, ":", rep("PhysActive", each = reps)))

# Plot regression coefficient estimates and error bars
plot_regr_coefs(regr_coefs = regr_coefs, res = res, cov_labels = cov_labels)

}
\seealso{
\code{\link[=get_regr_coefs]{get_regr_coefs()}} \code{\link[=plot_pattern_probs]{plot_pattern_probs()}} \code{\link[=plot_clus_dist]{plot_clus_dist()}}
\code{\link[=plot_pattern_profiles]{plot_pattern_profiles()}}
}
