% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_theta_modes}
\alias{plot_theta_modes}
\title{Plot theta modal exposure categories for each latent class}
\usage{
plot_theta_modes(
  res,
  item_labels = NULL,
  item_title = "Item",
  categ_labels = NULL,
  categ_title = "Consumption Level",
  class_labels = NULL,
  class_title = "Dietary Pattern",
  ...
)
}
\arguments{
\item{res}{Output from \code{swolca()}, \code{solca()}, or \code{wolca()} containing at
least \code{estimates} (or \code{estimates_unadj}) and \code{data_vars}}

\item{item_labels}{String vector of names for the exposure items. Jx1. If
\code{NULL} (default), numbers from 1 to J are used.}

\item{item_title}{String specifying the title for the exposure items.
Default is \code{"Item"}.}

\item{categ_labels}{String vector of names for the item categories. Rx1. If
\code{NULL} (default), numbers from 1 to R are used.}

\item{categ_title}{String specifying the title for the item categories.
Default is \code{"Consumption Level"}.}

\item{class_labels}{String vector of names for the latent classes. Kx1.
If \code{NULL} (default), numbers from 1 to K are used, where K is the final
determined number of latent classes.}

\item{class_title}{String specifying the title for the latent classes.
Default is \code{"Dietary Pattern"}.}

\item{\dots}{Additional arguments passed}
}
\value{
Returns a \code{ggplot2} object displaying a heatmap of the latent class patterns.
}
\description{
\code{plot_theta_modes} plots a heatmap of the latent class patterns, where the
patterns are defined by the category with the highest probability
(i.e., the model category) for each exposure item.
}
\seealso{
\code{\link[=swolca]{swolca()}} \code{\link[=solca]{solca()}} \code{\link[=wolca]{wolca()}}
}
