% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{post_process}
\alias{post_process}
\title{Post-process posterior samples to relabel and reduce classes}
\usage{
post_process(MCMC_out, p, d, q)
}
\arguments{
\item{MCMC_out}{Output from `run_MCMC_Rcpp` containing `pi_MCMC`, 
`theta_MCMC`, `xi_MCMC`, `c_all_MCMC`, `z_all_MCMC`, and `loglik_MCMC`}

\item{q}{Number of regression covariates excluding class assignment}
}
\value{
Returns list `post_MCMC_out` containing:
\describe{
  \item{\code{K_med}}{Median, across iterations, of number of classes with at least 5 percent of individuals}
  \item{\code{pi}}{Matrix of reduced and relabeled posterior samples for pi. (n_iter)x(K_med)}
  \item{\code{theta}}{Array of reduced and relabeled posterior samples for theta. (n_iter)xpx(K_med)xd}
  \item{\code{xi}}{Array of reduced and relabeled posterior samples for xi. (n_iter)x(K_med)xq}
  \item{\code{dendrogram}}{Hierarchical clustering dendrogram used for relabeling}
}
}
\description{
`post_process` conducts post-processing to cluster individuals into a 
reduced number of classes and reorder posterior parameter samples according 
to the reduced number of classes.
}
