% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_mod.R
\docType{class}
\name{Rcpp_rstantools_model_SWOLCA_main-class}
\alias{Rcpp_rstantools_model_SWOLCA_main-class}
\alias{Rcpp_rstantools_model_SWOLCA_main}
\title{S4 class for SWOLCA_main stan model}
\description{
S4 class for SWOLCA_main stan model inheriting \code{stanfit} class
}
\keyword{internal}
