% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_mod.R
\docType{class}
\name{Rcpp_rstantools_model_WOLCA_main-class}
\alias{Rcpp_rstantools_model_WOLCA_main-class}
\alias{Rcpp_rstantools_model_WOLCA_main}
\title{S4 class for WOLCA_main stan model}
\description{
S4 class for WOLCA_main stan model inheriting \code{stanfit} class
}
\keyword{internal}
