% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet_cpp}
\alias{rdirichlet_cpp}
\title{Draw from a Dirichlet distribution}
\usage{
rdirichlet_cpp(alpha)
}
\arguments{
\item{alpha}{Vector of positive concentration parameters. Length determines
the number of dimensions for the Dirichlet distribution.}
}
\value{
Vector result of a single draw from the specified Dirichlet distribution
}
\description{
Draw from a Dirichlet distribution
}
\keyword{internal}
