% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{reorder_classes}
\alias{reorder_classes}
\title{Reorder latent classes}
\usage{
reorder_classes(res, new_order)
}
\arguments{
\item{res}{An object of class \code{"swolca"}, \code{"solca"}, or \code{"wolca"}, resulting
from a call to \code{\link[=swolca]{swolca()}} \code{\link[=solca]{solca()}} or \code{\link[=wolca]{wolca()}}.}

\item{new_order}{Numeric vector specifying the new ordering of the classes.
For example, if there are three classes and the new desired order is to have
class 2 as the reference, followed by class 3 and class 1, then \code{new_order}
would be \code{c(2, 3, 1)}.}
}
\value{
Returns object \code{res_new} that is identical to input \code{res} but has updated
latent class ordering for \code{pi_red}, \code{theta_red}, \code{xi_red}, \code{pi_med},
\code{theta_med}, \code{xi_med}, and \code{c_all}.
}
\description{
\code{reorder_classes} changes the order of the latent classes to help with
plotting and to allow visualization of a different reference class for the
regression coefficients.
}
\details{
Latent class assignment variable \code{c_all} is also changed, so that if
\code{new_order = c(2, 3, 1)}, then \code{c_all == 2} will become \code{c_all == 1},
\code{c_all = 3} will become \code{c_all = 2}, and \code{c_all == 1} will become \code{c_all = 3}.
}
\examples{
# Load NHANES data
data(run_nhanes_swolca_results)
# Reorder latent classes
res_new <- reorder_classes(res = run_nhanes_swolca_results, 
                           new_order = c(3, 2, 5, 4, 1))
# Get posterior estimates for xi with class 3 as the reference
get_regr_coefs(res = res_new, ci_level = 0.95, digits = 2)                           

}
\seealso{
\code{\link[=plot_Phi_line]{plot_Phi_line()}} \code{\link[=plot_pi_boxplots]{plot_pi_boxplots()}}
}
