% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtruncnorm_cpp}
\alias{rtruncnorm_cpp}
\title{Draw from a truncated Normal distribution}
\usage{
rtruncnorm_cpp(n, a, b, mean, sd)
}
\arguments{
\item{n}{Number of draws}

\item{a}{Lower bound}

\item{b}{Upper bound}

\item{mean}{Mean}

\item{sd}{Standard deviation}
}
\value{
Result of a single draw from the specified truncated Normal distribution
}
\description{
Draw from a truncated Normal distribution
}
\keyword{internal}
