% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCMC_Rcpp.R
\name{run_MCMC_Rcpp}
\alias{run_MCMC_Rcpp}
\title{Run MCMC to get posterior samples}
\usage{
run_MCMC_Rcpp(
  OLCA_params,
  probit_params,
  n_runs,
  burn,
  thin,
  K,
  p,
  d,
  n,
  q,
  w_all,
  x_mat,
  y_all,
  V,
  alpha,
  eta,
  mu0,
  Sig0
)
}
\arguments{
\item{OLCA_params}{Output list from `init_OLCA()` containing `pi`, `c_all`, 
and `theta`}

\item{probit_params}{Output list from `init_probit()` containing `xi` and `z_all`}

\item{n_runs}{Number of MCMC iterations}

\item{burn}{Burn-in period}

\item{thin}{Thinning factor}

\item{K}{Number of classes}

\item{n}{Number of individuals}

\item{q}{Number of regression covariates excluding class assignment}

\item{w_all}{Weights normalized to sum to n. nx1}

\item{x_mat}{Categorical exposure matrix. nxp}

\item{y_all}{Vector of outcomes. nx1}

\item{V}{Regression design matrix without class assignment. nxq}

\item{alpha}{Vector of hyperparameters for pi. Kx1}

\item{eta}{Vector of hyperparameters for theta. dx1}

\item{mu0}{List of vectors of mean hyperparameters for xi. List of K qx1 vectors}

\item{Sig0}{List of matrices of variance hyperparameters for xi. List of K qxq matrices}
}
\value{
Returns list `MCMC_out` containing:
\describe{
  \item{\code{pi_MCMC}}{Matrix of posterior samples for pi. (n_iter)xK}
  \item{\code{theta_MCMC}}{Array of posterior samples for theta. (n_iter)xpxKxd}
  \item{\code{xi_MCMC}}{Array of posterior samples for xi. (n_iter)xKxq}
  \item{\code{c_all_MCMC}}{Matrix of posterior samples for c_all. (n_iter)xn}
  \item{\code{z_all_MCMC}}{Matrix of posterior samples for z_all. (n_iter)xn}
  \item{\code{loglik_MCMC}}{Vector of posterior samples for log-likelihood. (n_iter)x1}
}
}
\description{
`run_MCMC_Rcpp` runs the Gibbs sampler MCMC algorithm using Rcpp to obtain 
posterior samples.
}
