% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{run_nhanes_swolca_results}
\alias{run_nhanes_swolca_results}
\title{NHANES SWOLCA results}
\format{
List \code{res} containing:
\describe{
\item{\code{estimates_unadj}}{List of unadjusted posterior model results}
\item{\code{runtime}}{Total runtime for model}
\item{\code{data_vars}}{List of data variables used}
\item{\code{MCMC_out}}{List of full MCMC output}
\item{\code{post_MCMC_out}}{List of MCMC output after relabeling}
\item{\code{K_fixed}}{Number of classes used for the fixed sampler}
\item{\code{estimates}}{List of adjusted posterior model results with
correct uncertainty estimation}
}
}
\source{
Result from running \code{swolca()} on \code{data_nhanes}
}
\usage{
data(run_nhanes_swolca_results)
}
\description{
Results from running \code{swolca()} on the \code{data_nhanes} dataset
}
\examples{
data(run_nhanes_swolca_results)
}
\keyword{datasets}
