% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated data}
\format{
A list with 23 elements:
\describe{
\item{\code{samp_ind}}{Vector of population indices for sampled
individuals. nx1, where n=800.}
\item{\code{sample_wt}}{Vector of sampling weights for sampled
individuals. nx1}
\item{\code{N}}{Population size}
\item{\code{J}}{Number of exposure items}
\item{\code{R}}{Number of exposure categories}
\item{\code{H}}{Number of stratum (i.e., levels of S)}
\item{\code{N_s}}{Vector of population sizes for levels of S. Hx1, where H = 2.}
\item{\code{true_K}}{True number of latent classes, K}
\item{\code{true_Ai}}{\code{NULL} or vector of additional continuous variable
for sampled individuals. nx1}
\item{\code{true_Bi}}{\code{NULL} or vector of additional binary variable for
sampled individuals. nx1}
\item{\code{true_Si}}{Vector of true stratum indicators for sampled
individuals. nx1}
\item{\code{true_Ci}}{Vector of true latent class indicators for sampled
individuals. nx1}
\item{\code{true_pi}}{Vector of true pi values overall in the population.
Kx1, where K = 3.}
\item{\code{true_pi_s}}{\code{NULL} or HxK matrix of true pi values within each
level of S}
\item{\code{X_data}}{Matrix of multivariate categorical exposure for
sampled individuals. nxJ, where J = 20.}
\item{\code{true_global_patterns}}{Matrix of true global exposure patterns
defined by modal category. JxK}
\item{\code{true_global_thetas}}{Array of true thetas. JxKxR, where R = 4.}
\item{\code{Y_data}}{Vector of binary outcome for sampled individuals. nx1}
\item{\code{cluster_id}}{Vector of true cluster indicators for sampled
individuals. nx1}
\item{\code{cluster_size}}{Cluster size}
\item{\code{true_xi}}{Matrix of probit regression coefficients in mixture
reference coding. Kxq, where q is the number of covariate terms in the
regression, excluding all terms involving C; q = 2 for this dataset.}
\item{\code{true_Phi}}{Vector of true outcome probabilities for sampled
individuals. nx1}
\item{\code{true_Phi_mat}}{\code{NULL} or matrix of true outcome probabilities
for individuals aggregated by C and S. KxH}
}
}
\source{
Simulated data using \code{\link[=simulate_pop]{simulate_pop()}} and \code{\link[=simulate_samp]{simulate_samp()}}.
}
\usage{
data(sim_data)
}
\description{
A dataset simulated using \code{\link[=simulate_pop]{simulate_pop()}} and \code{\link[=simulate_samp]{simulate_samp()}}.
}
\examples{
data(sim_data)
x_mat <- sim_data$X_data
}
\keyword{datasets}
