% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_samp}
\alias{simulate_samp}
\title{Create and save simulated sample data}
\usage{
simulate_samp(
  sim_pop,
  samp_prop = 0.05,
  samp_size = NULL,
  strat = TRUE,
  strat_dist = c(0.5, 0.5),
  clust = TRUE,
  samp_seed = 101,
  save_res = TRUE,
  save_path
)
}
\arguments{
\item{sim_pop}{Simulated population output from \code{sim_pop()} function}

\item{samp_prop}{Proportion of population to sample. Default is 0.05.
Either \code{samp_prop} or \code{samp_size} must be specified.}

\item{samp_size}{Sample size. Default is \code{NULL}. Either \code{samp_prop} or
\code{samp_size} must be specified.}

\item{strat}{Boolean specifying whether to perform stratification by S_i.
Default is \code{TRUE}.}

\item{strat_dist}{Vector of relative stratum sizes in the sample. Components
must be proportions that sum to 1. Default is \code{c(0.5, 0.5)}, which results in
equal stratum sizes in the sample.}

\item{clust}{Boolean specifying whether to perform cluster sampling. Default
is \code{TRUE}.}

\item{samp_seed}{Numeric seed for population generation. Default is 1.}

\item{save_res}{Boolean specifying if results should be saved. Default = \code{TRUE}.}

\item{save_path}{String specifying directory and file name to save results,
e.g., "~/Documents/stratified_classes". Default is \code{NULL}.}
}
\value{
Returns list \code{sim_samp} containing:
\describe{
\item{\code{samp_ind}}{Vector of population indices for sampled
individuals. nx1}
\item{\code{sample_wt}}{Vector of sampling weights for sampled
individuals. nx1}
\item{\code{N}}{Population size}
\item{\code{J}}{Number of exposure items}
\item{\code{R}}{Number of exposure categories}
\item{\code{S}}{Number of stratum (i.e., levels of S_i)}
\item{\code{N_s}}{Vector of population sizes for levels of S_i. Sx1}
\item{\code{true_K}}{True number of latent classes, K}
\item{\code{true_Ai}}{\code{NULL} or vector of additional continuous variable
for sampled individuals. nx1}
\item{\code{true_Bi}}{\code{NULL} or vector of additional binary variable for
sampled individuals. nx1}
\item{\code{true_Si}}{Vector of true stratum indicators for sampled
individuals. nx1}
\item{\code{true_Ci}}{Vector of true latent class indicators for sampled
individuals. nx1}
\item{\code{true_pi}}{Vector of true pi values overall in the population. Kx1}
\item{\code{true_pi_s}}{\code{NULL} or SxK matrix of true pi values within each
level of S_i}
\item{\code{X_data}}{Matrix of multivariate categorical exposure for
sampled individuals. nxJ}
\item{\code{true_global_patterns}}{Matrix of true global exposure patterns
defined by modal category. JxK}
\item{\code{true_global_thetas}}{Array of true thetas. JxKxR}
\item{\code{Y_data}}{Vector of binary outcome for sampled individuals. nx1}
\item{\code{cluster_id}}{Vector of true cluster indicators for sampled
individuals. nx1}
\item{\code{cluster_size}}{Cluster size}
\item{\code{true_xi}}{Matrix of probit regression coefficients in mixture
reference coding. KxQ, where Q is the number of covariate terms in the
regression, excluding all terms involving C_i}
\item{\code{true_Phi}}{Vector of true outcome probabilities for sampled
individuals. nx1}
\item{\code{true_Phi_mat}}{\code{NULL} or matrix of true outcome probabilities
for individuals aggregated by C_i and S_i. KxS}
}
If \code{save_res = TRUE} (default), also saves \code{sim_samp} as
\verb{[save_path]_sim_samp.RData}.
}
\description{
\code{simulate_samp} creates and saves simulated sample data by sampling from the
simulated population according to input specifications.
}
\examples{
\dontrun{
### Example 1: Default values
# Create population
sim_pop <- simulate_pop(save_res = FALSE)
sim_samp <- simulate_samp(sim_pop = sim_pop, samp_prop = 0.05, strat = TRUE,
                          strat_dist = c(0.5, 0.5), clust = TRUE,
                          samp_seed = 101, save_res = FALSE)
}
}
\seealso{
\code{\link[=simulate_pop]{simulate_pop()}}
}
