% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_mod.R
\name{stanmodels}
\alias{stanmodels}
\alias{get_stanmodels}
\title{Get all stan models}
\usage{
get_stanmodels()
}
\value{
Returns \code{stanmodels} object
}
\description{
Get all stan models available in the package.
}
\keyword{internal}
