% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summarize_res}
\alias{summarize_res}
\title{Obtains table of model estimates}
\usage{
summarize_res(res, ci_level = 0.95, digits = 2)
}
\arguments{
\item{res}{An object of class \code{"swolca"} or \code{"wolca"}, resulting from a call
to \code{\link[=swolca]{swolca()}}, \code{\link[=swolca_var_adjust]{swolca_var_adjust()}}, \code{\link[=wolca]{wolca()}}, or \code{\link[=wolca_var_adjust]{wolca_var_adjust()}}.}

\item{ci_level}{Numeric from 0 to 1 specifying the credible interval level.
Default is 0.95, which gives a 95\\% equal-tailed interval composed of the
2.5\\% and 97.5\\% quantiles. For \code{wolca} objects, this must match the
\code{ci_level} parameter in the main function.}

\item{digits}{Integer indicating the number of decimal places to be used.
Default is 2, which rounds to the nearest hundredth.}
}
\value{
Returns dataframe \code{estimates_df} of the coefficients estimates and their
lower and upper bounds using the \code{ci_level} specified, rounded to the number
of digits specified in \code{digits}. \eqn{\pi} (class membership probability)
parameters are listed first, followed by \eqn{\theta} (item level probability)
parameters and \eqn{\xi} (regression coefficient) parameters.
}
\description{
\code{summarize_res} produces a summary table of the model estimates.
}
\examples{
data(run_nhanes_swolca_results)
estimates_df <- summarize_res(res = run_nhanes_swolca_results, ci_level = 0.95, 
                              digits = 2)

}
\seealso{
\code{\link[=get_regr_coefs]{get_regr_coefs()}}
}
