% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swolca.R
\name{swolca}
\alias{swolca}
\title{Run the SWOLCA model}
\usage{
swolca(
  x_mat,
  y_all,
  sampling_wt,
  cluster_id,
  stratum_id,
  V,
  K_max = 30,
  adapt_seed = NULL,
  class_cutoff = 0.05,
  alpha_adapt = NULL,
  eta_adapt = NULL,
  mu0_adapt = NULL,
  Sig0_adapt = NULL,
  alpha_fixed = NULL,
  eta_fixed = NULL,
  mu0_fixed = NULL,
  Sig0_fixed = NULL,
  K_true = NULL,
  fixed_seed = NULL,
  n_runs = 20000,
  burn = 10000,
  thin = 5,
  mod_stan,
  save_res = TRUE,
  save_path = NULL
)
}
\arguments{
\item{x_mat}{Categorical exposure matrix. nxp}

\item{y_all}{Vector of outcomes. nx1}

\item{sampling_wt}{Vector of survey sampling weights. nx1}

\item{cluster_id}{Vector of individual cluster IDs. nx1}

\item{stratum_id}{Vector of individual stratum IDs. nx1}

\item{V}{Regression design matrix without class assignment. nxq}

\item{K_max}{Upper limit for number of classes. Default is 30.}

\item{adapt_seed}{Numeric seed for adaptive sampler. Default is `NULL`.}

\item{class_cutoff}{Minimum class size proportion when determining number of
classes in adaptive sampler. Default is 0.05.}

\item{alpha_adapt}{Adaptive sampler hyperparameter for prior for pi. Default is `NULL`.}

\item{eta_adapt}{Adaptive sampler hyperparameter for prior for theta. Default is `NULL`.}

\item{mu0_adapt}{List of K qx1 vectors of adaptive sampler mean hyperparameters for xi. Default is `NULL`.}

\item{Sig0_adapt}{List of K qxq matrices of adaptive sampler variance hyperparameters for xi. Default is `NULL`.}

\item{alpha_fixed}{Fixed sampler hyperparameter for prior for pi. Default is `NULL`.}

\item{eta_fixed}{Fixed sampler hyperparameter for prior for theta. Default is `NULL`.}

\item{mu0_fixed}{List of K qx1 vectors of fixed sampler mean hyperparameters for xi. Default is `NULL`.}

\item{Sig0_fixed}{List of K qxq matrices of fixed sampler variance hyperparameters for xi. Default is `NULL`.}

\item{K_true}{True number of classes, if known. If `NULL` (default),
adaptive sampler is run. Otherwise, model uses fixed sampler directly.}

\item{fixed_seed}{Numeric seed for fixed sampler. Default is `NULL`.}

\item{n_runs}{Number of MCMC iterations. Default is 20000.}

\item{burn}{Number of MCMC iterations to drop as a burn-in period. Default is 10000.}

\item{thin}{Thinning factor for MCMC iterations. Default is 5.}

\item{mod_stan}{Stan model to be used for variance adjustment}

\item{save_res}{Boolean specifying if results should be saved. Default = `TRUE`.}

\item{save_path}{String specifying directory and file name to save results. Default is `NULL`.}
}
\value{
Saves and returns list `res` containing:
\describe{
  \item{\code{estimates_adj}}{List of adjusted posterior model results}
  \item{\code{runtime}}{Total runtime for model}
  \item{\code{data_vars}}{List of data variables used}
  \item{\code{MCMC_out}}{List of full MCMC output}
  \item{\code{post_MCMC_out}}{List of MCMC output after relabeling}
  \item{\code{K_MCMC}}{Adaptive sampler MCMC output for K}
}

Also saves list `adapt_MCMC` containing:
\describe{
  \item{\code{MCMC_out}}{List of full MCMC output}
  \item{\code{K_fixed}}{Number of classes to use for fixed sampler; output
  from adaptive sampler}
  \item{\code{K_MCMC}}{Adaptive sampler MCMC output for K}
}
}
\description{
`swolca` runs a supervised weighted overfitted latent class estimates (SWOLCA)
and saves and returns the results.
}
\details{
If hyperparameters are left as `NULL` (default), the following values are 
used.
}
\examples{
# Load data and obtain relevant variables
data("sim_data")
data_vars <- sim_data
x_mat <- data_vars$X_data            # Categorical exposure matrix, nxp
y_all <- c(data_vars$Y_data)         # Binary outcome vector, nx1
cluster_id <- data_vars$cluster_id  # Cluster indicators, nx1
sampling_wt <- data_vars$sample_wt
# Probit model includes C and S: C + S + C:S
stratum_id <- data_vars[[covs]]  # Stratifying variable, nx1
# Regression design matrix without class assignment, nxq
V_data <- data.frame(s = as.factor(stratum_id))
V <- model.matrix(~ s, V_data)
q <- ncol(V)  # Number of regression covariates excluding class assignment

# Stan model
mod_stan <- stanmodels$WSOLCA_main
# Run swolca
swolca(x_mat = x_mat, y_all = y_all, sampling_wt = sampling_wt, 
       cluster_id = cluster_id, stratum_id = stratum_id, V = V, adapt_seed = 1,
       n_runs = 200, burn = 100, thin = 5, mod_stan = mod_stan, save_res = FALSE)

}
