% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swolca_var_adjust.R
\name{unconstrain}
\alias{unconstrain}
\title{Unconstrain parameters}
\usage{
unconstrain(i, K, stan_model, pi, theta, xi)
}
\arguments{
\item{i}{Row index}

\item{K}{Number of classes}

\item{stan_model}{Stan model}

\item{pi}{MCMC matrix output for pi; MxK}

\item{theta}{MCMC array output for theta; MxJxKxR}

\item{xi:}{MCMC matrix output for xi; MxKxS}
}
\value{
Outputs vector of unconstrained parameters
}
\description{
\code{unconstrain} is a helper function that converts each row of an input array
of MCMC parameter output from constrained space to unconstrained space in Stan
}
\keyword{internal}
