% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolca_var_adjust.R
\name{unconstrain_wolca}
\alias{unconstrain_wolca}
\title{Unconstrain parameters for WOLCA}
\usage{
unconstrain_wolca(i, K, stan_model, pi, theta)
}
\arguments{
\item{i}{Row index}

\item{K}{Number of classes}

\item{stan_model}{Stan model}

\item{pi}{MCMC matrix output for pi; MxK}

\item{theta}{MCMC array output for theta; MxJxKxR}
}
\value{
Outputs vector of unconstrained parameters
}
\description{
\code{unconstrain_wolca} is a helper function that converts each row of an input array
of MCMC parameter output from constrained space to unconstrained space in Stan
}
\keyword{internal}
