% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_c}
\alias{update_c}
\title{Update c}
\usage{
update_c(c_all, n, K, J, theta, x_mat, pi, z_all, V, xi, y_all)
}
\arguments{
\item{n}{Number of individuals}

\item{K}{Number of classes}

\item{J}{Number of exposure items}

\item{theta}{Array parameter of item level probabilities. JxKxR}

\item{x_mat}{Categorical exposure matrix. nxp}

\item{pi}{Vector parameter of class membership probabilities. Kx1}

\item{z_all}{Vector of latent variables in the probit model. nx1}

\item{V}{Regression design matrix without class assignment. nxq}

\item{xi}{Matrix parameter xi for probit regression coefficients. Kxq}

\item{y_all}{Vector of outcomes. nx1}
}
\value{
Updated \code{c_all} vector after drawing from a Categorical distribution
with updated category event probabilities.
}
\description{
\code{update_c} updates the vector of individual class assignments, c, by drawing
from a Categorical distribution with updated category event probabilities.
}
\keyword{internal}
