% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_loglik}
\alias{update_loglik}
\title{Update individual log-likelihood}
\usage{
update_loglik(loglik, n, J, c_all, theta, x_mat, pi, z_all, V, xi, y_all)
}
\arguments{
\item{n}{Number of individuals}

\item{J}{Number of exposure items}

\item{theta}{Array parameter of item level probabilities. JxKxR}

\item{x_mat}{Categorical exposure matrix. nxp}

\item{pi}{Vector parameter of class membership probabilities. Kx1}

\item{z_all}{Vector of latent variables in the probit model. nx1}

\item{V}{Regression design matrix without class assignment. nxq}

\item{xi}{Matrix parameter xi for probit regression coefficients. Kxq}

\item{y_all}{Vector of outcomes. nx1}
}
\value{
Updated \code{loglik} vector after using the updated parameters and
latent variables.
}
\description{
\code{update_loglik} updates the vector of individual log-likelihoods using the
updated parameters and latent variables.
}
\keyword{internal}
