% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_loglik_wolca}
\alias{update_loglik_wolca}
\title{Update individual log-likelihood}
\usage{
update_loglik_wolca(loglik, n, J, c_all, theta, x_mat, pi)
}
\arguments{
\item{n}{Number of individuals}

\item{J}{Number of exposure items}

\item{theta}{Array parameter of item level probabilities. JxKxR}

\item{x_mat}{Matrix of multivariate categorical exposures. nxJ}

\item{pi}{Vector parameter of class membership probabilities. Kx1}
}
\value{
Updated \code{loglik} vector after using the updated parameters and
latent variables.
}
\description{
\code{update_loglik_wolca} updates the vector of individual log-likelihoods for
WOLCA using the updated parameters and latent variables.
}
\keyword{internal}
