% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_pi}
\alias{update_pi}
\title{Update pi}
\usage{
update_pi(pi, w_all, c_all, K, alpha)
}
\arguments{
\item{pi}{Vector parameter of class membership probabilities. Kx1}

\item{w_all}{Weights normalized to sum to n. nx1}

\item{K}{Number of classes}

\item{alpha}{Kx1 vector of hyperparameters for prior for class membership
probabilities \eqn{\pi}}
}
\value{
Updated \code{pi} vector after drawing from its posterior distribution
}
\description{
\code{update_pi} updates the vector parameter of class membership probabilities, pi, by
drawing from its posterior.
}
\keyword{internal}
