% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_theta}
\alias{update_theta}
\title{Update theta}
\usage{
update_theta(theta, p, K, d, eta, w_all, c_all, x_mat)
}
\arguments{
\item{theta}{Array parameter of item level probabilities. JxKxR}

\item{p}{Number of exposure items}

\item{K}{Number of classes}

\item{x_mat}{Categorical exposure matrix. nxp}
}
\value{
Updated `theta` array after drawing from its posterior distribution.
}
\description{
`update_theta` updates the array of item level probabilities, theta, by
drawing from its posterior.
}
\keyword{internal}
