% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_xi}
\alias{update_xi}
\title{Update xi}
\usage{
update_xi(xi, n, K, w_all, c_all, z_all, V, mu0, Sig0)
}
\arguments{
\item{xi}{Matrix parameter xi for probit regression coefficients. Kxq}

\item{n}{Number of individuals}

\item{K}{Number of classes}

\item{z_all}{Vector of latent variables in the probit model. nx1}

\item{V}{Regression design matrix without class assignment. nxq}
}
\value{
Updated \code{xi} matrix after drawing from its posterior distribution.
}
\description{
\code{update_xi} updates the matrix of outcome regression coefficients, xi, by
drawing from its posterior.
}
\keyword{internal}
