% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_z}
\alias{update_z}
\title{Update z}
\usage{
update_z(z_all, n, V, xi, c_all, y_all)
}
\arguments{
\item{z_all}{Vector of latent variables in the probit model. nx1}

\item{n}{Number of individuals}

\item{V}{Regression design matrix without class assignment. nxq}

\item{xi}{Matrix parameter xi for probit regression coefficients. Kxq}

\item{y_all}{Vector of outcomes. nx1}
}
\value{
Updated \code{z_all} vector after drawing from a Truncated Normal
distribution with updated mean.
}
\description{
\code{update_z} updates the vector of latent probit variables, z, by drawing
from a Truncated Normal distribution with updated mean.
}
\keyword{internal}
