% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolca_svyglm.R
\name{wolca_svyglm}
\alias{wolca_svyglm}
\title{Fit survey-weighted probit model for WOLCA}
\usage{
wolca_svyglm(
  res,
  y_all,
  glm_form,
  ci_level = 0.95,
  V_data = NULL,
  save_res = TRUE,
  save_path = NULL
)
}
\arguments{
\item{y_all}{Vector of binary outcomes. nx1}

\item{glm_form}{String specifying formula for probit regression, excluding
outcome and latent class. For example, \code{"~ 1"} for the model with only
latent class as covariates. All variables in \code{glm_form} must be found in \code{V_data}.
Do not specify interaction terms for latent class by additional covariates,
as these terms are already included.}

\item{ci_level}{Confidence interval level for probit regression coefficient
estimates. Default is \code{0.95}.}

\item{V_data}{Dataframe of additional regression covariates. nxq. Factor
covariates must be converted to factors. If \code{NULL} (default), no additional
covariates are to be included. All variables in \code{glm_form} must
be found in \code{V_data}.}

\item{save_res}{Boolean specifying if results should be saved. Default = \code{TRUE}.}

\item{save_path}{String specifying directory and file name to save results,
e.g., "~/Documents/run". Default is \code{NULL}.}
}
\value{
Returns an object \code{res} of class \code{"wolca"}, which includes all outputs from
\code{\link[=wolca]{wolca()}} as well as some updates. List \code{estimates} of \code{res} is updated to
contain the following additional objects:
\describe{
\item{\code{xi_est}}{Matrix of estimates for xi. (K_red)xq}
\item{\code{xi_est_lb}}{Matrix of confidence interval lower bound estimates for xi. (K_red)xq}
\item{\code{xi_est_ub}}{Matrix of confidence interval upper bound estimates for xi. (K_red)xq}
\item{\code{fit}}{\code{svyglm} class object with output from the \code{svyglm} regression model}
}
List \code{data_vars} of \code{res} is updated to contain the following additional objects:
\describe{
\item{\code{q}}{Number of regression covariates excluding class assignment.}
\item{\code{y_all}}{Vector of binary outcomes; nx1.}
\item{\code{V_data}}{Dataframe of additional regression covariates; nxq or \code{NULL}.}
\item{\code{V}}{Regression design matrix without class assignment; nxq.}
\item{\code{glm_form}}{String specifying formula for probit regression,
excluding outcome and latent class.}
\item{\code{ci_level}}{Confidence interval level for probit regression
coefficient estimates.}
}
The \code{runtime} output for \code{res} is also updated to include the runtime for the
probit regression in addition to the runtime for the main \code{wolca()} model.

If \code{save_res = TRUE} (default), the updated \code{res} object is saved as
\verb{[save_path]_wolca_results.RData}, overwriting the unsupervised results if
the file names are the same.
}
\description{
\code{wolca_svyglm} relates latent classes patterns derived from \code{wolca()} to a
binary outcome by fitting a survey-weighted probit model.
}
\details{
\code{wolca_svyglm} is the second step of a two-step approach that runs an
unsupervised WOLCA in the first step to derive latent class patterns and
subsequently treats the class assignments as fixed and includes them as
covariates in a frequentist survey-weighted probit regression model that uses
an asymptotic sandwich estimator for variance estimation.

\code{wolca_svyglm} specifies survey design and fits a survey-weighted probit
regression model according to the formula specified in \code{glm_form} using the
\code{svyglm()} function from the \code{survey} package (Lumley, 2023). Regression
coefficients and their confidence intervals are obtained from the \code{svyglm()}
output. If the residual degrees of freedom is less than 1, a Wald confidence
interval is manually calculated using a t-distribution with degrees of
freedom from the survey design. The point and interval estimates are then
converted into the mixture reference coding format to match the output format
from \code{\link[=swolca]{swolca()}}. \code{V_data} includes all covariates to include in the probit
regression other than latent class.

To save results, set \code{save_res = TRUE} (default) and \code{save_path} to a string
that specifies both the location and the beginning of the file name
(e.g., "~/Documents/run"). The file name will have "_wolca_results.RData"
appended to it, overwriting the unsupervised results if the file names are the
same.
}
\examples{
# Load data and obtain relevant variables
data("sim_data")
data_vars <- sim_data
x_mat <- data_vars$X_data            # Categorical exposure matrix, nxJ
y_all <- c(data_vars$Y_data)         # Binary outcome vector, nx1
cluster_id <- data_vars$cluster_id   # Cluster indicators, nx1
stratum_id <- data_vars$true_Si      # Stratum indicators, nx1
sampling_wt <- data_vars$sample_wt   # Survey sampling weights, nx1
n <- dim(x_mat)[1]                   # Number of individuals

# Probit model only includes latent class
V_data <- NULL # Additional regression covariates
# Survey-weighted regression formula
glm_form <- "~ 1"

# Run wolca
res <- wolca(x_mat = x_mat, sampling_wt = sampling_wt, 
             cluster_id = cluster_id, stratum_id = stratum_id, 
             run_sampler = "both", adapt_seed = 1, n_runs = 50, burn = 25, 
             thin = 1, save_res = FALSE)
       
# Apply variance adjustment to posterior estimates
res_adjust <- wolca_var_adjust(res = res, num_reps = 100, save_res = FALSE, 
                               adjust_seed = 1) 

# Run weighted outcome regression model
res_svyglm <- wolca_svyglm(res = res_adjust, y_all = y_all, 
                           glm_form = glm_form, ci_level = 0.95, 
                           V_data = V_data, save_res = FALSE)

}
\references{
Lumley T (2023). “survey: analysis of complex survey samples.” R
package version 4.2.
}
\seealso{
\code{\link[=wolca]{wolca()}}
}
