% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{catch_errors}
\alias{catch_errors}
\title{Catch errors in input variables for external functions}
\usage{
catch_errors(
  x_mat = NULL,
  y_all = NULL,
  sampling_wt = NULL,
  cluster_id = NULL,
  stratum_id = NULL,
  V_data = NULL,
  run_sampler = NULL,
  glm_form = NULL,
  Q = NULL,
  K_max = NULL,
  adapt_seed = NULL,
  fixed_seed = NULL,
  class_cutoff = NULL,
  alpha_adapt = NULL,
  eta_adapt = NULL,
  mu0_adapt = NULL,
  Sig0_adapt = NULL,
  K_fixed = NULL,
  alpha_fixed = NULL,
  eta_fixed = NULL,
  mu0_fixed = NULL,
  Sig0_fixed = NULL,
  n_runs = NULL,
  burn = NULL,
  thin = NULL,
  update = NULL,
  adjust_var = NULL,
  num_reps = NULL,
  save_res = NULL,
  save_path = NULL
)
}
\arguments{
\item{x_mat}{Matrix of multivariate categorical exposures. nxJ}

\item{y_all}{Vector of binary outcomes. nx1}

\item{sampling_wt}{Vector of survey sampling weights. nx1. Default is \code{NULL},
indicating no sampling weights and setting all weights to 1.}

\item{cluster_id}{Vector of individual cluster IDs. nx1. Default is \code{NULL},
indicating each individual is their own cluster.}

\item{stratum_id}{Vector of individual stratum IDs. nx1. Default is \code{NULL},
indicating no stratification.}

\item{V_data}{Dataframe of additional regression covariates. nxQ. Factor
covariates must be converted to factors. If \code{NULL} (default), no additional
covariates are to be included. All variables in \code{glm_form} must
be found in \code{V_data}.}

\item{run_sampler}{String specifying which sampler(s) should be run. Must be
one of \code{"both"} (default), \code{"fixed"}, or \code{"adapt"}. See Details.}

\item{glm_form}{String specifying formula for probit regression, excluding
outcome and latent class. For example, \code{"~ 1"} for the model with only
latent class as covariates. All variables in \code{glm_form} must be found in \code{V_data}.
Do not specify interaction terms for latent class by additional covariates,
as these terms are already included.}

\item{K_max}{Upper limit for number of classes. Default is 30.}

\item{adapt_seed}{Numeric seed for adaptive sampler. Default is \code{NULL}.}

\item{fixed_seed}{Numeric seed for fixed sampler. Default is \code{NULL}.}

\item{class_cutoff}{Minimum class size proportion when determining number of
classes. Default is 0.05.}

\item{alpha_adapt}{Adaptive sampler hyperparameter for prior for class
membership probabilities \eqn{\pi}. Default is \code{NULL} and default values are
used (see Details). If specified, must be (\code{K_max})x1.}

\item{eta_adapt}{Adaptive sampler hyperparameter for prior for item
level probabilities \eqn{\theta_{jk\cdot}} for each item \eqn{j} and class
\eqn{k}, assumed to be the same across classes. Default is \code{NULL} and default
values are used (see Details). If specified, must be JxR, where J is the
number of exposure items and R is the maximum number of categories for the exposure.}

\item{mu0_adapt}{Adaptive sampler mean hyperparameters for regression
coefficients \eqn{\xi_{k\cdot}} for each class \eqn{k}. Default is \code{NULL} and
default values are used (see Details). If specified, must be a list of \code{K_max}
vectors of dimension Qx1, where Q is the number of regression covariates
excluding latent class assignment.}

\item{Sig0_adapt}{Adaptive sampler variance hyperparameters for regression
coefficients \eqn{\xi_{k\cdot}} for each class \eqn{k}. Default is \code{NULL} and
default values are used (see Details). If specified, must be a list of \code{K_max}
QxQ matrices, where Q is the number of regression covariates excluding latent
class assignment.}

\item{K_fixed}{True number of classes, if known. Default is \code{NULL}, as it is
not necessary for the adaptive sampler. If bypassing the adaptive sampler and
running the fixed sampler directly, need to specify a value here. See Details.}

\item{alpha_fixed}{Fixed sampler hyperparameter for prior for \eqn{\pi}. Default is
\code{NULL} and default values are used (see Details). If specified, must be (\code{K_fixed})x1.}

\item{eta_fixed}{Fixed sampler hyperparameter for prior for item level
probabilities \eqn{\theta_{jk\cdot}} for each item \eqn{j} and class \eqn{k},
assumed to be the same across classes. Default is \code{NULL} and default values
are used (see Details). If specified, must be JxR, where J is the number of
exposure items and R is the maximum number of categories for the exposure.}

\item{mu0_fixed}{Fixed sampler mean hyperparameters for regression coefficients
\eqn{\xi_{k\cdot}} for each class \eqn{k}. Default is \code{NULL} and default
values are used (see Details). If specified, must be a list of \code{K_fixed}
vectors of dimension Qx1, where Q is the number of regression covariates
excluding latent class assignment.}

\item{Sig0_fixed}{Fixed sampler variance hyperparameters for regression
coefficients \eqn{\xi_{k\cdot}} for each class \eqn{k}. Default is \code{NULL} and
default values are used (see Details). If specified, must be a list of
\code{K_fixed} QxQ matrices, where Q is the number of regression covariates
excluding latent class assignment.}

\item{n_runs}{Number of MCMC iterations. Default is 20000.}

\item{burn}{Number of MCMC iterations to drop as a burn-in period. Default is 10000.}

\item{thin}{Thinning factor for MCMC iterations. Default is 5.}

\item{update}{Number specifying that MCMC progress updates should be printed
every \code{update} iterations. Default is 10.}

\item{save_res}{Boolean specifying if results should be saved. Default = \code{TRUE}.}

\item{save_path}{String specifying directory and file name to save results,
e.g., "~/Documents/run". Default is \code{NULL}.}

\item{model}{String specifying which model is used. Must be one of \code{swolca}
(default) or \code{wolca}}
}
\value{
The function stops and an error message is displayed if the input
variables are not acceptable
}
\description{
Catch input errors in variables necessary for package functions. All
parameters are set to \code{NULL} by default so that error checks
are only performed on relevant variables.
}
\details{
All parameters are set to \code{NULL} by default so that error checks
are only performed on relevant variables.
}
\keyword{internal}
