% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_mix_to_ref}
\alias{convert_mix_to_ref}
\title{Convert from mixture reference coding to reference cell coding}
\usage{
convert_mix_to_ref(est_xi)
}
\arguments{
\item{est_xi}{Matrix of xi parameter estimates in mixture reference coding. KxQ}
}
\value{
Returns vector \code{est_beta} of the probit regression coefficients
converted into reference cell coding with interactions. (K*Q)x1
}
\description{
Convert regression estimates \eqn{\xi} from mixture reference coding, a
combination of factor variable and reference cell coding, to standard
reference cell coding.
}
\keyword{internal}
