% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_probs}
\alias{convert_to_probs}
\title{Convert from mixture reference coding to P(Y=1|-) conditional probabilities}
\usage{
convert_to_probs(est_xi, glm_form, V, cov_name)
}
\arguments{
\item{est_xi}{Matrix of xi parameter estimates. Kxq}

\item{glm_form}{String specifying formula for probit regression, excluding
outcome and latent class. For example, \code{"~ 1"} for the model with only
latent class as covariates. All variables in \code{glm_form} must be found in \code{V_data}.
Do not specify interaction terms for latent class by additional covariates,
as these terms are already included.}

\item{cov_name}{String specifying name of covariate of interest. Must be
included in \code{glm_form}.}
}
\value{
Returns dataframe \code{probs} of the converted probabilities for the
covariate specified in \code{cov_name}, with number of rows equal to K and number
of columns equal to the number of categories for the covariate (including the
baseline category) plus one. The first column specifies the latent class,
the second column corresponds to the baseline category intercept for all K
latent classes, and the remaining columns correspond to the other categories
for the covariate.
}
\description{
Convert regression estimates \eqn{\xi} from mixture reference coding to
conditional probit regression probabilities, P(Y=1|-), for a given covariate.
}
\keyword{internal}
