% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{get_betas_c}
\alias{get_betas_c}
\title{Obtain betas matrix for generating categorical latent class assignment}
\usage{
get_betas_c(pi_mat, design_mat)
}
\arguments{
\item{pi_mat}{Matrix of class assignment probabilities, pi. SxK}

\item{design_mat}{Design matrix containing s_all}
}
\value{
Returns \code{beta_mat} matrix of betas to be used in a multinomial logistic
regression to generate a categorical variable C. \code{beta_mat} has R rows and
number of columns equal to the number of columns in the design matrix
}
\description{
Obtain matrix of betas that can be used to generate the categorical latent
class assignment variable C, depending on binary variable S, using the
specified matrix of assignment probabilities, \eqn{\pi}, and multinomial
logistic regression formula \eqn{C ~ S}
}
\examples{
pi_mat <- matrix(c(0.3, 0.5, 0.2, 
                   0.1, 0.6, 0.3), byrow = TRUE, nrow = 2)
formula_lc <- "~ s_all"
V <- data.frame(s_all = as.factor(c(rep(0, times = 60000),
                                    rep(1, times = 20000))))
design_mat <- model.matrix(as.formula(formula_lc), data = V)
get_betas_c(pi_mat = pi_mat, design_mat = design_mat)
}
\keyword{internal}
