% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{get_levels}
\alias{get_levels}
\title{Get levels of a variable}
\usage{
get_levels(df, var)
}
\arguments{
\item{df}{Dataframe including variable of interest}

\item{var}{String specifying variable of interest}
}
\value{
Outputs string vector \code{levels} containing the levels of the variable
of interest. If the variable is not a factor, outputs the empty string \code{""}.
}
\description{
\code{convert_ref_to_mix} converts from reference cell coding to a combination of
factor variable and reference cell coding, referred to as mixture reference
coding.
}
\seealso{
\code{\link[=vars_across_clus]{vars_across_clus()}}
}
\keyword{internal}
