% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_regr_coefs}
\alias{get_regr_coefs}
\title{Obtains table of regression coefficients}
\usage{
get_regr_coefs(res, ci_level = 0.95, digits = 2)
}
\arguments{
\item{res}{An object of class \code{"swolca"}, \code{"solca"}, or \code{"wolca"}, resulting
from a call to \code{\link[=swolca]{swolca()}} \code{\link[=solca]{solca()}} or \code{\link[=wolca]{wolca()}}.}

\item{ci_level}{Numeric from 0 to 1 specifying the credible interval level.
Default is 0.95, which gives a 95\\% equal-tailed interval composed of the
2.5\\% and 97.5\\% quantiles. For \code{wolca()} results, this must match the
\code{ci_level} parameter in the main function.}

\item{digits}{Integer indicating the number of decimal places to be used.
Default is 2, which rounds to the nearest hundredth.}
}
\value{
Returns a character vector of the source code for regression coefficients
table using the format specified in \code{format}.
}
\description{
\code{get_regr_coefs} produces a summary table of the regression coefficients,
converted to standard reference cell coding.
}
\examples{
data(run_nhanes_swolca_results)
get_regr_coefs(res = run_nhanes_swolca_results, ci_level = 0.95, digits = 2)

}
\seealso{
\code{\link[=plot_Phi_line]{plot_Phi_line()}}
}
