% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regr_coefs.R
\name{get_regr_coefs}
\alias{get_regr_coefs}
\title{Obtains table of regression coefficients}
\usage{
get_regr_coefs(res, ci_level = 0.95, digits = 2)
}
\arguments{
\item{res}{Output from \code{swolca()}, \code{solca()}, or \code{wolca()} containing at
least \code{estimates} (or \code{estimates_unadj}) and \code{data_vars}}

\item{ci_level}{Numeric from 0 to 1 specifying the credible interval level.
Default is 0.95, which gives a 95\\% equal-tailed interval composed of the
2.5\\% and 97.5\\% quantiles. For \code{wolca()} results, this must match the
\code{ci_level} parameter in the main function.}

\item{digits}{Integer indicating the number of decimal places to be used.
Default is 2, which rounds to the nearest hundredth.}
}
\value{
Returns a character vector of the source code for regression coefficients
table using the format specified in \code{format}.
}
\description{
\code{get_regr_coefs} produces a summary table of the regression coefficients,
converted to standard reference cell coding.
}
\seealso{
\code{\link[=swolca]{swolca()}} \code{\link[=solca]{solca()}} \code{\link[=wolca]{wolca()}}
}
