% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{init_OLCA}
\alias{init_OLCA}
\title{Initialize OLCA model}
\usage{
init_OLCA(K, n, J, R, alpha, eta)
}
\arguments{
\item{K}{Number of classes}

\item{n}{Number of individuals}

\item{J}{Number of exposure items}

\item{R}{Maximum number of exposure categories}

\item{alpha}{Kx1 vector of hyperparameters for prior for class membership
probabilities \eqn{\pi}}

\item{eta}{JxR matrix of hyperparameters for prior for item consumption
level probabilities \eqn{\theta_{jk\cdot}} for each item \eqn{j} and class
\eqn{k}, assumed to be the same across classes.}
}
\value{
Returns list \code{OLCA_params} containing:
\describe{
\item{\code{pi}}{Vector parameter pi for class membership probabilities. Kx1}
\item{\code{theta}}{Array parameter theta for item category probabilities. JxKxR}
\item{\code{c_all}}{Vector of random initial class assignments. nx1}
}
}
\description{
\code{init_OLCA} initializes priors and variables for an overfitted latent class
analysis (OLCA) given hyperparameters.
}
\details{
First, class membership probabilities \eqn{\pi} are initialized by drawing
from a Dirichlet distribution with hyperparameter \eqn{\alpha = 1/K} for each
of the K components. Then, class assignments \code{c_all} are initialized for each
individual by drawing from a Categorical distribution with parameter \eqn{\pi}.
Finally, item consumption level probabilities \eqn{\theta} are initialized by
drawing from a Dirichlet distribution with hyperparameter \eqn{\eta},
independently for each exposure item and latent class.
}
\examples{
K <- 30; n <- 4000; J <- 30; R <- 4
alpha <- rep(1, K) / K
eta <- matrix(1, nrow = J, ncol = R)
OLCA_params <- init_OLCA(K = K, n = n, J = J, R = R, alpha = alpha, eta = eta)
# OLCA_params

}
\seealso{
\code{\link[=init_probit]{init_probit()}} \code{\link[=swolca]{swolca()}} \code{\link[=solca]{solca()}} \code{\link[=wolca]{wolca()}}
}
