% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_probit.R
\name{init_probit}
\alias{init_probit}
\title{Initialize probit model}
\usage{
init_probit(K, n, q, V, mu0, Sig0, y_all, c_all)
}
\arguments{
\item{K}{Number of classes}

\item{n}{Number of individuals}

\item{q}{Number of regression covariates excluding class assignment}

\item{V}{Regression design matrix without class assignment. nxq}

\item{mu0}{List of vectors of mean hyperparameters for xi. List of K qx1 vectors}

\item{Sig0}{List of matrices of variance hyperparameters for xi. List of K qxq matrices}

\item{y_all}{Vector of outcomes. nx1}

\item{c_all}{Vector of random initial class assignments. nx1}
}
\value{
Returns list `probit_params` containing:
\describe{
  \item{\code{xi}}{Matrix parameter xi for probit regression coefficients. Kxq}
  \item{\code{z_all}}{Vector of latent variables in the probit model. nx1}
}
}
\description{
`init_probit` initializes priors and variables for the probit regression model 
given hyperparameters.
}
\examples{
# Load data and obtain relevant variables
data("sim_data")
data_vars <- sim_data
x_mat <- data_vars$X_data            # Categorical exposure matrix, nxp
y_all <- c(data_vars$Y_data)         # Binary outcome vector, nx1
n <- dim(x_mat)[1]        # Number of individuals
p <- dim(x_mat)[2]        # Number of exposure items
d <- max(apply(x_mat, 2,  # Number of exposure categories
function(x) length(unique(x))))  

# Probit model covariates only include latent class 
V <- matrix(1, nrow = n)
q <- ncol(V)

# Set hyperparameters
K <- 30
alpha <- rep(1, K) / K
eta <- rep(1, d)
mu0 <- Sig0 <- vector("list", K)
for (k in 1:K) {
  # MVN(0,1) hyperprior for prior mean of xi
  mu0[[k]] <- stats::rnorm(n = q)
  # InvGamma(3.5, 6.25) hyperprior for prior variance of xi. Assume uncorrelated
  # components and mean variance 2.5 for a weakly informative prior on xi
  Sig0[[k]] <- diag(LaplacesDemon::rinvgamma(n = q, shape = 3.5, scale = 6.25), 
  nrow = q, ncol = q)
}

# First initialize OLCA params
OLCA_params <- init_OLCA(K = K, n = n, p = p, d = d, alpha = alpha, eta = eta)

# Initialize probit model
probit_params <- init_probit(K = K, n = n, q = q, V = V, mu0 = mu0, 
Sig0 = Sig0, y_all = y_all, c_all = OLCA_params$c_all)
# probit_params

}
