% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_Phi_line}
\alias{plot_Phi_line}
\title{Plot conditional probability of outcome for each latent class for a specified
categorical covariate.}
\usage{
plot_Phi_line(
  res,
  cov_name,
  ci_level = NULL,
  cov_labels = NULL,
  class_labels = NULL,
  class_title = "Dietary Pattern",
  x_title = NULL,
  y_title = "Probability of Outcome",
  ...
)
}
\arguments{
\item{res}{An object of class \code{"swolca"}, \code{"solca"}, or \code{"wolca"}, resulting
from a call to \code{\link[=swolca]{swolca()}} \code{\link[=solca]{solca()}} or \code{\link[=wolca]{wolca()}}.}

\item{cov_name}{String specifying the covariate to plot}

\item{ci_level}{Numeric from 0 to 1 specifying the credible interval level.
If \code{NULL} (default), no credible intervals are included in the plot.
Specifying 0.95 gives a 95\\% equal-tailed interval composed of the 2.5\\% and
97.5\\% quantiles. For \code{wolca()} results, this must match the \code{ci_level}
parameter in the main function.}

\item{cov_labels}{String vector specifying the category labels for the
covariate of interest. Must be the same length as the number of categories
in the covariate specified by \code{cov_name}. If \code{NULL} (default), numbers from 1
to the number of covariate categories are used.}

\item{class_labels}{String vector of names for the latent classes. Kx1.
If \code{NULL} (default), numbers from 1 to K are used, where K is the final
determined number of latent classes.}

\item{class_title}{String specifying the title for the latent classes.
Default is \code{"Dietary Pattern"}.}

\item{x_title}{String specifying x-axix label. If \code{NULL} (default), \code{cov_name}
is used}

\item{y_title}{String specifying the title for the y-axis. Default is
\code{"Probability of Outcome"}.}

\item{\dots}{Additional arguments passed}
}
\value{
Returns a \code{ggplot2} object displaying a grouped line plot of the conditional
probability of the outcome across categories of the covariate, for each
latent class.
}
\description{
\code{plot_Phi_line} plots a grouped line plot of the conditional probability of
the outcome, obtained by transforming the probit regression coefficients to
the probability scale, for a categorical covariate, with each line
corresponding to a latent class.
}
\examples{
data(run_nhanes_swolca_results)

# Default labels
plot_Phi_line(res = run_nhanes_swolca_results, cov_name = "racethnic")

# Specifying labels
cov_labels <- c("NH White", "NH Black", "NH Asian", "Hispanic/Latino", 
                "Other/Mixed")
class_labels <- paste0("Class", 1:5)
x_title <- "Race and Ethnicity"
plot_Phi_line(res = run_nhanes_swolca_results, cov_name = "racethnic",
             cov_labels = cov_labels, class_labels = class_labels, 
             x_title = x_title)
}
\seealso{
\code{\link[=get_regr_coefs]{get_regr_coefs()}} \code{\link[=plot_theta_probs]{plot_theta_probs()}} \code{\link[=plot_pi_boxplots]{plot_pi_boxplots()}} \code{\link[=plot_theta_modes]{plot_theta_modes()}}
}
