% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_theta_modes}
\alias{plot_theta_modes}
\title{Plot theta modal exposure categories for each latent class}
\usage{
plot_theta_modes(
  res,
  item_labels = NULL,
  item_title = "Item",
  categ_labels = NULL,
  categ_title = "Consumption Level",
  class_labels = NULL,
  class_title = "Dietary Pattern",
  ...
)
}
\arguments{
\item{res}{Output from \code{swolca()}, \code{solca()}, or \code{wolca()} containing at
least \code{estimates} (or \code{estimates_unadj}) and \code{data_vars}}

\item{item_labels}{String vector of names for the exposure items. Jx1. If
\code{NULL} (default), numbers from 1 to J are used.}

\item{item_title}{String specifying the title for the exposure items.
Default is \code{"Item"}.}

\item{categ_labels}{String vector of names for the item categories. Rx1. If
\code{NULL} (default), numbers from 1 to R are used.}

\item{categ_title}{String specifying the title for the item categories.
Default is \code{"Consumption Level"}.}

\item{class_labels}{String vector of names for the latent classes. Kx1.
If \code{NULL} (default), numbers from 1 to K are used, where K is the final
determined number of latent classes.}

\item{class_title}{String specifying the title for the latent classes.
Default is \code{"Dietary Pattern"}.}

\item{\dots}{Additional arguments passed}
}
\value{
Returns a \code{ggplot2} object displaying a heatmap of the latent class patterns.
}
\description{
\code{plot_theta_modes} plots a heatmap of the latent class patterns, where the
patterns are defined by the category with the highest probability
(i.e., the model category) for each exposure item.
}
\examples{
data(run_nhanes_swolca_results)

# Default labels
plot_theta_modes(res = run_nhanes_swolca_results)

# Specifying labels
item_labels <- c("Citrus/Melon/Berries", "Other Fruits", "Fruit Juice", 
                  "Dark Green Vegs", "Tomatoes", "Oth Red/Orange Vegs",
                  "Potatoes", "Other Starchy Vegs", "Other Vegetables",
                  "Whole Grains", "Refined Grains", "Meat", "Cured Meats",
                  "Organ Meat", "Poultry", "Seafood (High n-3)", "Seafood (Low n-3)",
                  "Eggs", "Soybean Products", "Nuts and Seeds", "Legumes (Protein)",
                  "Milk", "Yogurt", "Cheese", "Oils", "Solid Fats", "Added Sugar",
                  "Alcoholic Drinks")
categ_labels <- c("None", "Low", "Med", "High")
class_labels <- paste0("Class ", 1:5)
plot_theta_modes(res = run_nhanes_swolca_results, item_labels = item_labels,
                 categ_labels = categ_labels, class_labels = class_labels)

}
\seealso{
\code{\link[=plot_theta_probs]{plot_theta_probs()}} \code{\link[=plot_pi_boxplots]{plot_pi_boxplots()}} \code{\link[=plot_Phi_line]{plot_Phi_line()}}
}
