% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_theta_probs}
\alias{plot_theta_probs}
\title{Plot probabilities of exposure categories, theta, for each latent class}
\usage{
plot_theta_probs(
  res,
  item_labels = NULL,
  categ_labels = NULL,
  categ_title = "Consumption Level",
  class_labels = NULL,
  class_title = "Dietary Pattern",
  y_title = "Consumption Level Probability",
  ...
)
}
\arguments{
\item{res}{Output from \code{swolca()}, \code{solca()}, or \code{wolca()} containing at
least \code{estimates} (or \code{estimates_unadj}) and \code{data_vars}}

\item{item_labels}{String vector of names for the exposure items. Jx1. If
\code{NULL} (default), numbers from 1 to J are used.}

\item{categ_labels}{String vector of names for the item categories. Rx1. If
\code{NULL} (default), numbers from 1 to R are used.}

\item{categ_title}{String specifying the title for the item categories.
Default is \code{"Consumption Level"}.}

\item{class_labels}{String vector of names for the latent classes. Kx1.
If \code{NULL} (default), numbers from 1 to K are used, where K is the final
determined number of latent classes.}

\item{class_title}{String specifying the title for the latent classes.
Default is \code{"Dietary Pattern"}.}

\item{y_title}{String specifying the title for the y-axis. Default is
\code{"Consumption Level Probability"}.}

\item{\dots}{Additional arguments passed}
}
\value{
Returns a \code{ggplot2} object displaying a grouped barplot of the probability of
the exposure categories, for each exposure item and each latent class
}
\description{
\code{plot_theta_probs} plots a grouped barplot of the probability of the exposure
categories, for each exposure item and each latent class.
}
\examples{
data(run_nhanes_swolca_results)
plot_theta_probs(res = run_nhanes_swolca_results)

}
\seealso{
\code{\link[=plot_theta_modes]{plot_theta_modes()}} \code{\link[=plot_pi_boxplots]{plot_pi_boxplots()}} \code{\link[=plot_Phi_line]{plot_Phi_line()}}
}
