% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_wolca.R
\name{post_process_wolca}
\alias{post_process_wolca}
\title{Post-process posterior samples to relabel and reduce classes for WOLCA}
\usage{
post_process_wolca(MCMC_out, p, d)
}
\arguments{
\item{MCMC_out}{Output from `run_MCMC_Rcpp_wolca` containing `pi_MCMC`, 
`theta_MCMC`, and `c_all_MCMC`,}

\item{p}{Number of exposure items}

\item{d}{Number of exposure categories}
}
\value{
Returns list `post_MCMC_out` containing:
\describe{
  \item{\code{K_med}}{Median, across iterations, of number of classes with at least 5 percent of individuals}
  \item{\code{pi}}{Matrix of reduced and relabeled posterior samples for pi. (n_iter)x(K_med)}
  \item{\code{theta}}{Array of reduced and relabeled posterior samples for theta. (n_iter)xpx(K_med)xd}
  \item{\code{dendrogram}}{Hierarchical clustering dendrogram used for relabeling}
}
}
\description{
`post_process_wolca` conducts post-processing to cluster individuals into a 
reduced number of classes and reorder posterior parameter samples according 
to the reduced number of classes for the unsupervised WOLCA model.
}
\examples{
# Load data and obtain relevant variables
data("sim_data")
data_vars <- sim_data
x_mat <- data_vars$X_data            # Categorical exposure matrix, nxp
cluster_id <- data_vars$cluster_id  # Cluster indicators, nx1
sampling_wt <- data_vars$sample_wt

# Obtain dimensions
n <- dim(x_mat)[1]        # Number of individuals
p <- dim(x_mat)[2]        # Number of exposure items
d <- max(apply(x_mat, 2,  # Number of exposure categories
function(x) length(unique(x))))  
# Obtain normalized weights
kappa <- sum(sampling_wt) / n   # Weights norm. constant
w_all <- c(sampling_wt / kappa) # Weights normalized to sum to n, nx1

# Set hyperparameters for fixed sampler
K <- 3
alpha <- rep(1, K) / K
eta <- rep(1, d)

# First initialize OLCA params
OLCA_params <- init_OLCA(K = K, n = n, p = p, d = d, alpha = alpha, eta = eta)

# Then run MCMC sampling
MCMC_out <- run_MCMC_Rcpp_wolca(OLCA_params = OLCA_params, n_runs = 50, 
burn = 25, thin = 5, K = K, p = p, d = d, n = n, w_all = w_all, x_mat = x_mat, 
alpha = alpha, eta = eta)

# Then run post-process relabeling
post_MCMC_out <- post_process_wolca(MCMC_out = MCMC_out, p = p, d = d)
# post_MCMC_out
# plot(post_MCMC_out$dendrogram)

}
