% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_mod.R
\docType{data}
\name{rstantools_model_SWOLCA_main}
\alias{rstantools_model_SWOLCA_main}
\title{Get SWOLCA_main stan model}
\format{
An object of class \code{C++Class} of length 1.
}
\usage{
rstantools_model_SWOLCA_main
}
\arguments{
\item{stan_name}{Name of stan model file without file extension}
}
\value{
Returns `mod_stan` containing the specified stan model
}
\description{
Get stan model used in the `swolca()` function for the variance adjustment.
}
\keyword{datasets}
