% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_c_wolca}
\alias{update_c_wolca}
\title{Update c for WOLCA}
\usage{
update_c_wolca(c_all, n, K, J, theta, x_mat, pi)
}
\arguments{
\item{n}{Number of individuals}

\item{K}{Number of classes}

\item{J}{Number of exposure items}

\item{theta}{Array parameter of item level probabilities. JxKxR}

\item{x_mat}{Categorical exposure matrix. nxJ}

\item{pi}{Vector parameter of class membership probabilities. Kx1}
}
\value{
Updated \code{c_all} vector after drawing from a Categorical distribution
with updated category event probabilities for the unsupervised model.
}
\description{
Updates the vector of individual class assignments, c, for the unsupervised
model.
}
\keyword{internal}
