% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_adjust.R
\name{var_adjust}
\alias{var_adjust}
\title{Post-processing variance adjustment}
\usage{
var_adjust(
  mod_stan,
  estimates,
  K,
  p,
  d,
  n,
  q,
  x_mat,
  y_all,
  V,
  w_all,
  stratum_id,
  cluster_id
)
}
\arguments{
\item{mod_stan}{Stan model}

\item{estimates}{Output from `get_estimates()` containing `K_red`, `pi_red`, 
`theta_red`, `xi_red`, `pi_med`, `theta_med`, `xi_med`, `Phi_med`, `c_all`, 
`pred_class_probs`, `loglik_med`}

\item{K}{Number of classes}

\item{n}{Number of individuals}

\item{q}{Number of regression covariates excluding class assignment}

\item{x_mat}{Categorical exposure matrix. nxp}

\item{y_all}{Vector of outcomes. nx1}

\item{V}{Regression design matrix without class assignment. nxq}

\item{w_all}{Weights normalized to sum to n. nx1}

\item{stratum_id}{Vector of stratifying variable for individuals. nx1}

\item{cluster_id}{Vector of cluster indicators. nx1}
}
\value{
Returns list `estimates_adj` containing:

\describe{
  \item{\code{pi_red_adj}}{Matrix of adjusted posterior samples for pi. Mx(K_red)}
  \item{\code{theta_red_adj}}{Array of adjusted posterior samples for theta. Mxpx(K_red)xd}
  \item{\code{xi_red_adj}}{Array of adjusted posterior samples for xi. Mx(K_red)xq}
  \item{\code{pi_med_adj}}{Vector of adjusted posterior median estimates for pi. (K_red)x1}
  \item{\code{theta_med_adj}}{Array of adjusted posterior median estimates for theta. px(K_red)xd}
  \item{\code{xi_med_adj}}{Matrix of adjusted posterior median estimates for xi. (K_red)xq}
  \item{\code{Phi_med_adj}}{Vector of adjusted individual outcome probabilities. nx1}
  \item{\code{c_all}}{Vector of final individual class assignments from `get_estimates()`. nx1}
  \item{\code{pred_class_probs}}{Matrix of individual posterior class probabilities from `get_estimates()`. nx(K_red)}
  \item{\code{loglik_med}}{Vector of final indiviudal log-likehoods from `get_estimates()`. nx1} 
}
}
\description{
`var_adj` applies applies the post-processing variance adjustment
}
