% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolca.R
\name{wolca}
\alias{wolca}
\title{Run the WOLCA model}
\usage{
wolca(
  x_mat,
  sampling_wt,
  cluster_id,
  stratum_id,
  run_sampler = "both",
  K_max = 30,
  adapt_seed = NULL,
  class_cutoff = 0.05,
  alpha_adapt = NULL,
  eta_adapt = NULL,
  alpha_fixed = NULL,
  eta_fixed = NULL,
  K_fixed = NULL,
  fixed_seed = NULL,
  n_runs = 20000,
  burn = 10000,
  thin = 5,
  save_res = TRUE,
  save_path = NULL
)
}
\arguments{
\item{x_mat}{Matrix of multivariate categorical exposures. nxJ}

\item{sampling_wt}{Vector of survey sampling weights. nx1. Default is \code{NULL},
indicating no sampling weights and setting all weights to 1.}

\item{cluster_id}{Vector of individual cluster IDs. nx1. Default is \code{NULL},
indicating each individual is their own cluster.}

\item{stratum_id}{Vector of individual stratum IDs. nx1. Default is \code{NULL},
indicating no stratification.}

\item{run_sampler}{String specifying which sampler(s) should be run. Must be
one of \code{"both"} (default), \code{"fixed"}, or \code{"adapt"}.}

\item{K_max}{Upper limit for number of classes. Default is 30.}

\item{adapt_seed}{Numeric seed for adaptive sampler. Default is \code{NULL}.}

\item{class_cutoff}{Minimum class size proportion when determining number of
classes in adaptive sampler. Default is 0.05.}

\item{alpha_adapt}{Adaptive sampler hyperparameter for prior for class
membership probabilities \eqn{\pi}. Default is \code{NULL} and default values are
used. If specified, must be (\code{K_max})x1.}

\item{eta_adapt}{Adaptive sampler hyperparameter for prior for item consumption
level probabilities \eqn{\theta_{jk\cdot}} for each item \eqn{j} and class
\eqn{k}, assumed to be the same across classes. Default is \code{NULL} and default
values are used. If specified, must be JxR, where J is the number of exposure
items and R is the maximum number of categories for the exposure.}

\item{alpha_fixed}{Fixed sampler hyperparameter for prior for \eqn{\pi}. Default is
\code{NULL} and default values are used. If specified, must be (\code{K_fixed})x1.}

\item{eta_fixed}{Fixed sampler hyperparameter for prior for item consumption
level probabilities \eqn{\theta_{jk\cdot}} for each item \eqn{j} and class
\eqn{k}, assumed to be the same across classes. Default is \code{NULL} and default
values are used. If specified, must be JxR, where J is the number of exposure
items and R is the maximum number of categories for the exposure.}

\item{K_fixed}{True number of classes, if known. Default is \code{NULL}, as it is
not necessary for the adaptive sampler. If bypassing the adaptive sampler and
running the fixed sampler directly, need to specify a value here.}

\item{fixed_seed}{Numeric seed for fixed sampler. Default is \code{NULL}.}

\item{n_runs}{Number of MCMC iterations. Default is 20000.}

\item{burn}{Number of MCMC iterations to drop as a burn-in period. Default is 10000.}

\item{thin}{Thinning factor for MCMC iterations. Default is 5.}

\item{save_res}{Boolean specifying if results should be saved. Default = \code{TRUE}.}

\item{save_path}{String specifying directory and file name to save results,
e.g., "~/Documents/run". Default is \code{NULL}.}
}
\value{
If the fixed sampler is run, returns an object \code{res} of class \code{"wolca"}; a
list containing the following:
\describe{
\item{\code{estimates}}{List of posterior model results, resulting from a
call to \code{\link[=run_MCMC_Rcpp]{run_MCMC_Rcpp()}}}
\item{\code{runtime}}{Total runtime for model}
\item{\code{data_vars}}{List of data variables used, including:
\code{n}: Sample size.
\code{J}: Number of exposure items.
\code{R_j}: Number vector of number of exposure categories for each item; Jx1.
\code{R}: Maximum number of exposure categories across items.
\code{w_all}: Vector of sampling weights normalized to sum to n; nx1.
\code{sampling_wt}: Vector of survey sampling weights; nx1.
\code{x_mat}: Matrix of multivariate categorical exposures; nxJ.
\code{stratum_id}: Vector of individual stratum IDs; nx1 or NULL.
\code{cluster_id}: Vector of individual cluster IDs; nx1 or NULL.
}
\item{\code{MCMC_out}}{List of full MCMC output, resulting from a call to
\code{\link[=run_MCMC_Rcpp]{run_MCMC_Rcpp()}}}
\item{\code{post_MCMC_out}}{List of MCMC output after relabeling, resulting
from a call to \code{\link[=post_process]{post_process()}}}
\item{\code{K_fixed}}{Number of classes used for the fixed sampler}
}
If \code{save_res = TRUE} (default), also saves \code{res} as
\verb{[save_path]_wolca_results.RData}.

If only the adaptive sampler is run (i.e., \code{run_sampler} = \code{"adapt"}), returns
list \code{res} containing:
\describe{
\item{\code{MCMC_out}}{List of full MCMC output, resulting from a call to
\code{\link[=run_MCMC_Rcpp]{run_MCMC_Rcpp()}}}
\item{\code{K_fixed}}{Number of classes used for the fixed sampler,
obtained from the adaptive sampler}
\item{\code{K_MCMC}}{Adaptive sampler MCMC output for K; Mx1, where M is
the number of MCMC iterations after burn-in and thinning.}
}
If \code{save_res = TRUE} (default), also saves \code{res} as
\verb{[save_path]_wolca_adapt.RData}.
}
\description{
\code{wolca} runs an unsupervised weighted overfitted latent class analysis
(WOLCA) described in Wu et al. (2023) and Stephenson et al. (2023), then
saves and returns the results.
}
\details{
If no survey sample adjustments are desired, leave \code{sampling_wt}, \code{stratum_id},
and \code{cluster_id} to their default \code{NULL} values.

By default, the function will run two samplers: the adaptive sampler followed
by the fixed sampler. The adaptive sampler determines the number of latent
classes, which is then used in the fixed sampler for parameter estimation.
If the number of latent classes is already known and only the fixed sampler
needs to be run, specify \code{"fixed"} for the \code{run_sampler} argument and specify a
number for \code{K_fixed}. If only the adaptive sampler is to be run, specify
\code{"adapt"} for the \code{run_sampler} argument. Use \code{adapt_seed} (default is \code{NULL})
to specify a seed for the adaptive sampler, and use \code{fixed_seed} (default is
\code{NULL}) to specify a separate seed for the fixed sampler.

\code{x_mat} is an nxJ matrix with each row corresponding to the J-dimensional
categorical exposure for an individual. If there is no clustering present,
\code{cluster_id} should be set to the individual IDs.
\code{K_max} is the maximum number of latent classes allowable, to
be used for the overfitted latent class model if the adaptive sampler is run.
\code{class_cutoff} is the minimum size of each class as a proportion of the
population, used when determining the number of latent classes.

To save results, set \code{save_res = TRUE} (default) and \code{save_path} to a string
that specifies both the location and the beginning of the file name
(e.g., "~/Documents/run"). The file name will have "_wolca_adapt.RData" or
"_wolca_results.RData" appended to it.

If hyperparameters for the adaptive or fixed sampler are left as \code{NULL}
(default), the following default values are used. Let \eqn{K} refer to
\code{K_max} for the adaptive sampler and \code{K_fixed} for the fixed sampler.
For \eqn{\pi}, a Dirichlet prior with hyperparameter \eqn{\alpha = 1/K} for
each component. For \eqn{\theta_{jk\cdot}}, a Dirichlet prior with
hyperparameter  \eqn{\eta_j} equal to \code{rep(1, R_j)} where \code{R_j} is the number
of categories for exposure item j. If \code{R_j < R}, the remaining categories have
hyperparameter set to 0.01. This is done independently for each exposure item j
and is assumed to be the same across latent classes. Note that hyperparameters
for the fixed sampler should probably only be specified if running the
fixed sampler directly, bypassing the adaptive sampler.
}
\examples{
# Load data and obtain relevant variables
data("sim_data")
data_vars <- sim_data
x_mat <- data_vars$X_data            # Categorical exposure matrix, nxJ
cluster_id <- data_vars$cluster_id   # Cluster indicators, nx1
stratum_id <- data_vars$true_Si      # Stratum indicators, nx1
sampling_wt <- data_vars$sample_wt   # Survey sampling weights, nx1
n <- dim(x_mat)[1]                   # Number of individuals

# Run wolca
res <- wolca(x_mat = x_mat, sampling_wt = sampling_wt, 
             cluster_id = cluster_id, stratum_id = stratum_id, 
             run_sampler = "both", adapt_seed = 1, n_runs = 50, burn = 25, 
             thin = 1, save_res = FALSE)

}
\references{
Stephenson, B. J. K., Wu, S. M., Dominici, F. (2023). Identifying dietary
consumption patterns from survey data: a Bayesian nonparametric latent class
model. Journal of the Royal Statistical Society Series A: Statistics in
Society, qnad135.

Wu, S. M., Williams, M. R., Savitsky, T. D., & Stephenson, B. J.
(2023). Derivation of outcome-dependent dietary patterns for low-income women
obtained from survey data using a Supervised Weighted Overfitted Latent Class
Analysis. arXiv preprint arXiv:2310.01575.
}
